--- steering/Reference/steering-display/.gitignore
@@ -1,2 +1,0 @@
-.pio
-.vscode
