--- steering/Reference/steering-display/platformio.ini
@@ -1,41 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[platformio]
-default_envs = q23-steering
-
-[env]
-lib_extra_dirs =
-    ../../../cal
-lib_deps =
-    cal
-    https://github.com/RudolphRiedel/FT800-FT813/archive/refs/heads/5.x.zip
-framework = arduino
-build_flags = -D EVE_NHD_43 
-check_tool = cppcheck
-check_flags = cppcheck: --addon=cert.py --suppress=unusedFunction
-
-[env:q23-steering]
-platform = ststm32
-board = genericSTM32F446RE
-board_build.mcu = stm32f446re
-board_build.f_cpu = 180000000L
-monitor_speed = 115200
-debug_tool = blackmagic
-upload_protocol = blackmagic
-; upload_port = COM4
-; monitor_port = COM3
-build_flags = 
-    ${env.build_flags}
-         -D EVE_CS=PA4
-         -D EVE_PDN=8
-         -D EVE_SPI=SPI1
-    -DENABLE_HWSERIAL2
-
