--- steering/Reference/steering-display/src/tft_data.c
@@ -1,69 +1,0 @@
-
-#include <stdint.h>
-
-#if !defined(PROGMEM)
-    #define PROGMEM
-#endif
-
-const uint8_t logo[5298] PROGMEM =
-{
-    0x78, 0x9c, 0xed, 0x5d, 0x4d, 0x8c, 0x1c, 0xc5, 0x15, 0xe, 0x3f, 0x7, 0xef, 0x85, 0xf8, 0x80, 0x84, 0xc3, 0xcd, 0x83, 0xd, 0xcc, 0x44, 0x42, 0xac, 0x14, 0xe, 0x4e, 0x6e, 0x9e, 0xb, 0xde, 0x45, 0x42, 0x58, 0x4a, 0x24, 0x96, 0xdc, 0xa6, 0x4f, 0x1e, 0x2, 0xe, 0x1b, 0x81, 0xc4, 0x3a, 0xf2, 0x65, 0x46, 0x1a, 0x89, 0x89, 0xb4, 0xe0, 0x95, 0xb0, 0xc4, 0x38, 0xa, 0x87, 0x45, 0x22, 0x78, 0x23, 0x25, 0xca, 0x2, 0x91, 0xbc, 0x1c, 0x38, 0x6c, 0x24, 0x14, 0x7c, 0xb0, 0x94, 0x65, 0x31, 0xed, 0x71, 0xad, 0xd8, 0xc9, 0xb2, 0x20, 0x66, 0xa5, 0x59, 0x69, 0xc7, 0x68, 0xc0, 0x95, 0xae, 0x9e, 0x9e, 0x9e, 0xaa, 0x7a, 0xaf, 0x7e, 0xba, 0x7b, 0x7e, 0x76, 0x4d, 0xbf, 0x27, 0xaf, 0x67, 0xa6, 0x7b, 0x7a, 0xba, 0xeb, 0x7d, 0xf5, 0xea, 0xd5, 0xfb, 0xa9, 0xa2, 0xe7, 0xe9, 0x1d, 0xc7, 0x75, 0x5a, 0xa7, 0x3f, 0x2a, 0x1a, 0x7b, 0x8b, 0xdf, 0x51, 0xfc, 0x63, 0xa6, 0x71, 0xb7, 0xfd, 0x9d, 0xc1, 0x29, 0x31, 0x1a, 0xb7, 0x14, 0xe, 0x36, 0xa7, 0xd4, 0xa7, 0x71, 0xcb, 0xe2, 0x80, 0x72, 0xfd, 0xfc, 0xb8, 0x5, 0xb7, 0xef, 0x68, 0xec, 0x32, 0x39, 0xa0, 0x9c, 0x92, 0x4c, 0xe3, 0x96, 0xc8, 0xc1, 0x64, 0x2b, 0xaa, 0xd3, 0x5a, 0x67, 0xa6, 0xf3, 0xb3, 0xf6, 0xad, 0xbd, 0xb5, 0xbd, 0x85, 0xd6, 0xac, 0xc7, 0x93, 0xcd, 0xc9, 0xe6, 0xe1, 0x3b, 0x80, 0x1f, 0x6f, 0xee, 0x6c, 0xaf, 0xb5, 0x52, 0x24, 0x25, 0x66, 0x3, 0x7e, 0x66, 0xda, 0x6b, 0x7b, 0x93, 0xcd, 0xe7, 0x1b, 0x13, 0x8d, 0x55, 0x52, 0x26, 0x5, 0xf2, 0x90, 0x9b, 0xe1, 0x99, 0x64, 0xdc, 0x83, 0xcf, 0x77, 0xb9, 0x25, 0x77, 0xae, 0x93, 0x22, 0x29, 0x11, 0x2b, 0xe9, 0x6, 0x3d, 0xd2, 0x9e, 0x6c, 0x4e, 0x79, 0xf8, 0xc9, 0x7, 0x98, 0xc9, 0x13, 0x87, 0xf1, 0x66, 0x69, 0xb3, 0xbc, 0xf, 0x78, 0x90, 0x77, 0xe1, 0x6c, 0xe6, 0xdd, 0x13, 0xed, 0x54, 0x2b, 0x25, 0x62, 0x94, 0x56, 0xe8, 0x5a, 0xab, 0xea, 0x61, 0x88, 0x21, 0x88, 0xe1, 0xa7, 0x14, 0xc8, 0x2e, 0x60, 0xe2, 0xf1, 0xa6, 0x81, 0x75, 0x67, 0x60, 0xc7, 0x88, 0xd5, 0x37, 0x4d, 0x47, 0xf9, 0xfb, 0x83, 0x67, 0x12, 0xf5, 0x15, 0x32, 0xee, 0x72, 0x3a, 0xbe, 0x25, 0x61, 0x94, 0x4e, 0xb4, 0xff, 0xb6, 0xbd, 0xe8, 0xa1, 0xe8, 0x64, 0x80, 0x21, 0xa5, 0xec, 0x55, 0x52, 0x25, 0x1, 0x47, 0xc5, 0x45, 0xf0, 0xce, 0xd1, 0x21, 0x47, 0x85, 0x38, 0xd5, 0x3b, 0xf1, 0x2e, 0x9, 0xf7, 0x8f, 0xe3, 0xb2, 0x87, 0xa4, 0xc7, 0x9b, 0x29, 0x92, 0x12, 0x30, 0x42, 0x6b, 0xad, 0x5c, 0xc3, 0x1b, 0xcf, 0x3c, 0x14, 0x95, 0xd5, 0xfd, 0x9a, 0xef, 0xdf, 0xdc, 0x39, 0x8e, 0xaa, 0xef, 0xc3, 0x6f, 0x49, 0x12, 0x77, 0xa4, 0xdf, 0x71, 0xd4, 0xef, 0xf4, 0x1a, 0xd1, 0xe6, 0x18, 0xb8, 0x82, 0x87, 0x24, 0x52, 0x6c, 0xa4, 0x48, 0x4a, 0xc0, 0x80, 0xda, 0x7b, 0x87, 0x36, 0xd8, 0x88, 0x56, 0xea, 0x8d, 0x66, 0x8a, 0x5e, 0x2c, 0xc8, 0x44, 0xa7, 0x9b, 0xcc, 0xe3, 0x91, 0x2, 0x71, 0x8e, 0xea, 0x17, 0xf5, 0x57, 0xb3, 0xd3, 0x82, 0x0, 0x4b, 0x79, 0xf2, 0x15, 0xa9, 0xa7, 0x48, 0x8a, 0xcb, 0x80, 0xe6, 0x3a, 0xc5, 0x6, 0xc3, 0x51, 0x39, 0xae, 0x9c, 0xf0, 0xf3, 0xb1, 0x11, 0x49, 0xc4, 0x8f, 0xff, 0xce, 0x61, 0xba, 0x7, 0x8e, 0x70, 0x18, 0xd2, 0xe0, 0x9d, 0x44, 0xb9, 0x2f, 0xe4, 0x2a, 0x5, 0x2, 0x66, 0x6f, 0x74, 0xec, 0xf2, 0x39, 0x38, 0xc, 0xe8, 0x74, 0xb3, 0xe0, 0x8d, 0x6b, 0x5a, 0x9, 0x99, 0xb4, 0x90, 0x5a, 0xb, 0xe0, 0x57, 0x50, 0x22, 0xc0, 0xb1, 0xb9, 0xb6, 0x9, 0x5f, 0xaa, 0x33, 0xc1, 0x1d, 0x38, 0x9e, 0xa5, 0xd4, 0x4e, 0x6d, 0xee, 0xb8, 0x2c, 0x51, 0xad, 0x93, 0xf3, 0x34, 0x92, 0xc3, 0xb7, 0x34, 0x6f, 0x3b, 0x43, 0x6c, 0xd8, 0x59, 0xd5, 0xfc, 0xb7, 0x78, 0x4b, 0x45, 0xf5, 0x5d, 0x9d, 0x25, 0x6f, 0x77, 0x9e, 0xd, 0x96, 0xf8, 0xef, 0xfb, 0xaf, 0x32, 0xee, 0x6c, 0x6a, 0x73, 0xc7, 0x65, 0x89, 0x8e, 0xb4, 0x17, 0x3d, 0x8d, 0xa4, 0x98, 0xab, 0x99, 0x6c, 0x6f, 0x9d, 0xdc, 0xa2, 0x58, 0x4a, 0x76, 0x7a, 0x4d, 0xf7, 0xcb, 0xe8, 0x5c, 0x40, 0xf1, 0xad, 0xf0, 0x75, 0x79, 0xf3, 0xa8, 0xbb, 0x94, 0xda, 0xdc, 0x71, 0x59, 0xa2, 0x85, 0x56, 0xc6, 0x2d, 0xe0, 0x33, 0x36, 0xbe, 0xf, 0xdb, 0xf8, 0x69, 0x64, 0x94, 0xc5, 0xd1, 0x1e, 0x36, 0xe3, 0x19, 0xb4, 0x9f, 0x45, 0x4d, 0x6a, 0x63, 0x57, 0x91, 0x2e, 0x92, 0x32, 0x24, 0xbb, 0x91, 0x22, 0x29, 0x26, 0x4b, 0xf4, 0x4c, 0x33, 0x18, 0xdb, 0x14, 0x32, 0x74, 0xb0, 0x91, 0xe, 0x22, 0x2d, 0xea, 0x31, 0x41, 0xba, 0x8e, 0xe, 0x4b, 0x4, 0x3b, 0xd7, 0xd1, 0x5c, 0x4d, 0x8d, 0x1c, 0xe9, 0x8, 0x61, 0x36, 0xf7, 0xa2, 0x7b, 0x25, 0x45, 0x52, 0x5c, 0x16, 0xe8, 0xf0, 0xd7, 0x3e, 0x92, 0x84, 0x96, 0x76, 0xc4, 0x59, 0x94, 0xde, 0x32, 0x52, 0x59, 0x54, 0xea, 0x6f, 0xd8, 0xb0, 0xe2, 0x9b, 0x8, 0x82, 0xa0, 0x67, 0x4b, 0x7f, 0x2f, 0xfd, 0x79, 0xa2, 0xef, 0x75, 0x48, 0x23, 0x26, 0x9, 0x18, 0x43, 0x52, 0x12, 0xf9, 0xc6, 0xf5, 0x1, 0xe9, 0xcf, 0x8b, 0x86, 0xc2, 0xe8, 0xfe, 0x80, 0x40, 0x2b, 0xa5, 0x11, 0x93, 0x4, 0x2c, 0xd0, 0xe9, 0xa6, 0x80, 0x24, 0x4c, 0xfb, 0xa8, 0x91, 0x22, 0xdb, 0x26, 0x26, 0x3b, 0x1c, 0x9b, 0x8f, 0xcb, 0xc7, 0x9, 0x18, 0xef, 0xfc, 0xf8, 0xd, 0x7a, 0x3f, 0x26, 0x7d, 0xa4, 0x46, 0x69, 0x70, 0x84, 0x45, 0x4c, 0xe, 0x7f, 0x9d, 0x22, 0x29, 0x26, 0xb, 0xb4, 0xdc, 0x92, 0x46, 0xb7, 0x10, 0x13, 0x8, 0xbe, 0x70, 0x4, 0xa8, 0xe4, 0x9, 0x67, 0xff, 0xb8, 0x76, 0x88, 0x36, 0x4b, 0xb3, 0xc3, 0x8a, 0x99, 0xfd, 0xef, 0x32, 0x24, 0x4d, 0xa7, 0x36, 0x77, 0x5c, 0x16, 0xe8, 0x48, 0x3b, 0xc2, 0xe8, 0x66, 0x23, 0x3f, 0xfd, 0x5c, 0x4f, 0x85, 0x24, 0xc9, 0x4e, 0x4b, 0x88, 0x16, 0x62, 0xf8, 0x3e, 0xf7, 0x49, 0x9e, 0x6c, 0xa5, 0x11, 0x93, 0xf8, 0xcc, 0xd1, 0x15, 0xba, 0xe8, 0x47, 0xdc, 0x22, 0x60, 0xc2, 0x74, 0xcc, 0x24, 0x5f, 0xf3, 0x3c, 0xdf, 0x84, 0x3c, 0xd3, 0x67, 0xf2, 0x71, 0xcd, 0x39, 0xde, 0xc8, 0xe9, 0xce, 0x74, 0x52, 0xad, 0x14, 0x93, 0x5, 0xca, 0x6d, 0x64, 0x2, 0x7f, 0x92, 0x23, 0x4a, 0x80, 0x48, 0x72, 0xb0, 0x9f, 0x93, 0x41, 0xe9, 0x89, 0xc8, 0xb3, 0xf5, 0x68, 0x63, 0x28, 0x4c, 0x36, 0x33, 0x4, 0xbf, 0x9c, 0x46, 0x4c, 0x12, 0xb1, 0x40, 0x87, 0xbf, 0x3e, 0xea, 0x6a, 0xa2, 0xb7, 0x10, 0x21, 0xa6, 0xb9, 0x9a, 0xc9, 0x5b, 0xe0, 0xbd, 0x72, 0xf4, 0xb2, 0x36, 0xfa, 0x81, 0xd4, 0xc8, 0xd0, 0xde, 0x3b, 0xbc, 0xa2, 0x3f, 0x7b, 0x3b, 0x2d, 0x45, 0x4c, 0xea, 0xe3, 0x96, 0xcf, 0xc1, 0x61, 0x81, 0xd6, 0x3c, 0x9b, 0xdb, 0x52, 0x1a, 0x36, 0xd2, 0xb, 0x8f, 0xf1, 0xd9, 0xb2, 0xf2, 0x59, 0x2a, 0xe4, 0x5a, 0x22, 0x5a, 0xfb, 0xbb, 0x0, 0x2d, 0x28, 0x7e, 0x7a, 0x7f, 0x99, 0xcd, 0x9d, 0x46, 0x4c, 0x12, 0x30, 0x47, 0xaf, 0x76, 0x1c, 0xd7, 0xb1, 0xf6, 0x47, 0x2b, 0x6c, 0x15, 0xd2, 0x93, 0x8b, 0xef, 0xd7, 0x24, 0x79, 0x8f, 0x33, 0x64, 0xb8, 0x95, 0x3, 0xf, 0xe9, 0x35, 0xa9, 0x3c, 0x9e, 0xa, 0x47, 0x1d, 0xee, 0x75, 0x9e, 0x4c, 0x6c, 0xdc, 0x4c, 0x91, 0x14, 0x97, 0x79, 0x5d, 0x4e, 0xbf, 0x22, 0x79, 0x82, 0xb4, 0x7e, 0x44, 0x8d, 0x50, 0xf6, 0x11, 0xc4, 0x64, 0xec, 0x90, 0x55, 0x32, 0xd1, 0x98, 0x6e, 0x2c, 0x6d, 0xef, 0x7c, 0x7b, 0xba, 0x35, 0xdb, 0x5a, 0x68, 0xfd, 0x77, 0xef, 0xb3, 0xbd, 0xf6, 0xde, 0xad, 0x1, 0xf3, 0xf2, 0x5e, 0x99, 0x58, 0xf6, 0x1, 0x8d, 0x8f, 0x81, 0x45, 0x4c, 0x6a, 0x9d, 0x14, 0x49, 0x31, 0x59, 0xa0, 0x6a, 0x23, 0xe3, 0x4a, 0xd9, 0x92, 0xb2, 0xc6, 0x31, 0x48, 0x88, 0xa1, 0xe8, 0xa4, 0x87, 0xa2, 0x45, 0x32, 0xdd, 0x98, 0x6c, 0xde, 0xda, 0x9b, 0xeb, 0x8c, 0x66, 0x9d, 0x93, 0xe9, 0x46, 0xde, 0x30, 0x37, 0xd3, 0xdc, 0x77, 0x18, 0x37, 0xc9, 0xbb, 0x47, 0xd2, 0x88, 0x49, 0x5c, 0x16, 0x68, 0xb6, 0x69, 0x6d, 0x29, 0x29, 0xec, 0xd, 0x36, 0x3, 0x2a, 0xb8, 0xcf, 0x37, 0xd6, 0xf6, 0x56, 0x46, 0x82, 0xa0, 0x1e, 0x5d, 0xdd, 0xce, 0x88, 0x63, 0x9c, 0x10, 0x9, 0x51, 0xf4, 0x5, 0x80, 0xbb, 0x8c, 0xbb, 0x90, 0xce, 0xde, 0xe2, 0xb2, 0x40, 0x47, 0xda, 0xf9, 0x7e, 0xae, 0x9b, 0xce, 0x3, 0x23, 0x7b, 0xa, 0x36, 0xbb, 0x91, 0xd0, 0xb2, 0x67, 0x6b, 0x14, 0xdc, 0x9d, 0x6d, 0x90, 0xc9, 0x3a, 0x74, 0x5a, 0x10, 0x67, 0xb, 0x44, 0xb0,
-    0x8b, 0x2c, 0xfd, 0x5, 0x2c, 0x4b, 0x69, 0x67, 0x5b, 0xba, 0xf0, 0xb8, 0xe5, 0x73, 0x90, 0x98, 0xa3, 0x15, 0xba, 0xca, 0x7b, 0x27, 0xd, 0xfe, 0x41, 0x19, 0x4b, 0x5d, 0x8d, 0x34, 0xd5, 0x18, 0x3d, 0x8e, 0x58, 0x1f, 0x38, 0xa9, 0xf1, 0xd0, 0x3b, 0xf2, 0x73, 0xa0, 0xf6, 0x37, 0xcb, 0x52, 0x9a, 0x4e, 0x67, 0x6f, 0xf1, 0x59, 0xa0, 0xbe, 0x77, 0x52, 0xd6, 0x41, 0x5a, 0x2b, 0x29, 0xd0, 0x51, 0xe, 0x81, 0x3e, 0x99, 0xd1, 0x50, 0xad, 0xe7, 0xa1, 0x97, 0x33, 0x86, 0xf9, 0x48, 0xa1, 0xe6, 0xee, 0xbb, 0x7f, 0xd3, 0x88, 0x49, 0x22, 0x16, 0x88, 0xe5, 0x96, 0xa0, 0x73, 0x6a, 0x85, 0x24, 0x1c, 0xe1, 0x78, 0xc1, 0x1b, 0xdb, 0x80, 0x9f, 0x58, 0x45, 0xe7, 0xd9, 0x7a, 0x3b, 0x71, 0xf9, 0xc4, 0x2b, 0xf2, 0xe5, 0xb2, 0x1b, 0xca, 0xdc, 0x61, 0x5b, 0x2c, 0x79, 0x3d, 0xa1, 0x90, 0x46, 0x4c, 0xe2, 0xb3, 0x40, 0xa8, 0x77, 0x52, 0xf6, 0xc1, 0x10, 0xf0, 0x8e, 0x93, 0x4, 0x98, 0xfd, 0xc, 0x49, 0x22, 0xcf, 0xcd, 0x89, 0x97, 0x7, 0x36, 0x77, 0x70, 0x57, 0xe, 0xde, 0x13, 0xb0, 0x7c, 0x6, 0xc2, 0x46, 0xe7, 0x74, 0xdd, 0x92, 0x4, 0xcc, 0xd1, 0x5c, 0xd7, 0x3b, 0x69, 0xec, 0xbd, 0x25, 0x34, 0x9a, 0xe1, 0x6c, 0x3e, 0x84, 0xeb, 0xa4, 0xa1, 0xdf, 0x77, 0xd7, 0xe6, 0x2e, 0xeb, 0xe3, 0xb4, 0xb8, 0x67, 0x43, 0x78, 0x9d, 0x71, 0x9f, 0x49, 0x6b, 0x4c, 0xe2, 0x33, 0x47, 0x9c, 0x77, 0xd2, 0x8c, 0x25, 0xe4, 0x8, 0x92, 0x79, 0x48, 0x47, 0x71, 0xdf, 0xe1, 0xbc, 0x53, 0x85, 0x14, 0x75, 0xdc, 0x84, 0x3b, 0xc2, 0x22, 0x26, 0xd5, 0xd4, 0xe6, 0x8e, 0xcf, 0x2, 0x2d, 0x35, 0xee, 0xe, 0x46, 0x9, 0x7, 0xc3, 0x8a, 0x56, 0x5f, 0x8d, 0x7c, 0x1e, 0xcd, 0xd1, 0x15, 0xde, 0xe6, 0xc6, 0x75, 0x28, 0xff, 0x3f, 0xcc, 0x70, 0xf0, 0x39, 0x4f, 0xe, 0xa5, 0x11, 0x93, 0xf8, 0x2c, 0xd0, 0xac, 0x3e, 0x8e, 0xab, 0xb2, 0xbc, 0x7b, 0x51, 0x50, 0x32, 0x35, 0xca, 0xcc, 0x43, 0x81, 0x42, 0x9b, 0x5b, 0xe5, 0x37, 0x95, 0x7d, 0x4c, 0xc8, 0x68, 0x87, 0x44, 0x4c, 0x86, 0xfb, 0x4, 0x77, 0x16, 0xb, 0xf4, 0x33, 0xad, 0x67, 0x46, 0x33, 0x66, 0x10, 0x7c, 0x1e, 0x3d, 0xd4, 0xcc, 0xc, 0x81, 0xae, 0x6e, 0x1b, 0xb2, 0x62, 0xf0, 0xea, 0x17, 0x71, 0xfe, 0xe0, 0x8d, 0x6f, 0x69, 0xc4, 0x24, 0x1, 0x73, 0xb4, 0x42, 0x57, 0xc5, 0x38, 0x6e, 0x24, 0x1e, 0xf1, 0x3c, 0x5a, 0xa0, 0x5, 0x59, 0x9b, 0xda, 0xad, 0x96, 0x22, 0x71, 0x1a, 0x31, 0x49, 0xc4, 0x2, 0x4d, 0x35, 0x2, 0xef, 0xa4, 0x3c, 0xa, 0x40, 0xfb, 0xa3, 0xaf, 0x9b, 0xc2, 0x28, 0xa8, 0x3c, 0x8f, 0xae, 0xf, 0x53, 0x2b, 0x9, 0x24, 0xc5, 0x7a, 0x6c, 0xf5, 0xaa, 0xa0, 0x91, 0x18, 0x92, 0x64, 0x4b, 0x6f, 0xee, 0x95, 0xb1, 0xcb, 0xe7, 0xe0, 0xb0, 0x40, 0x87, 0x9b, 0xc8, 0x28, 0x21, 0x8f, 0xc, 0xaa, 0xec, 0x24, 0x64, 0x1e, 0x3d, 0xd7, 0x99, 0x9b, 0x1b, 0xc5, 0x9d, 0xd7, 0x3a, 0xef, 0xb8, 0xf1, 0xb5, 0x69, 0xa8, 0x93, 0x48, 0x4e, 0xb2, 0xf4, 0xea, 0x74, 0xf9, 0xcc, 0xd8, 0x25, 0x74, 0x50, 0x58, 0xa0, 0x35, 0x39, 0x1a, 0xaa, 0xc7, 0x8f, 0xc0, 0xd8, 0x3c, 0xba, 0x3e, 0xb2, 0xb5, 0xe3, 0xab, 0x8d, 0x9f, 0xa8, 0x2c, 0x25, 0x33, 0xb6, 0x82, 0xbe, 0x92, 0x27, 0xab, 0x64, 0x4, 0x79, 0xc, 0xe3, 0x96, 0xf8, 0x48, 0xb0, 0x34, 0xd7, 0x29, 0x89, 0xde, 0x49, 0x71, 0x6e, 0x6d, 0xd0, 0x4d, 0x9a, 0x79, 0xf4, 0x8, 0xa8, 0xd6, 0x99, 0xd8, 0x30, 0xc6, 0x7b, 0x74, 0x9e, 0x4b, 0xdf, 0xd2, 0x3, 0x75, 0xdd, 0xc3, 0xa1, 0x71, 0xcb, 0x7c, 0xe8, 0x48, 0xaa, 0xd3, 0x43, 0x1b, 0x5c, 0xde, 0x98, 0xc9, 0x4f, 0x29, 0xf9, 0x70, 0xc, 0xf3, 0xe8, 0x11, 0x50, 0x9d, 0xe6, 0x36, 0x84, 0xf1, 0x19, 0xcb, 0x91, 0x11, 0x5f, 0x71, 0xef, 0xd0, 0x88, 0xc9, 0xf0, 0x68, 0xdc, 0x72, 0x1f, 0x2a, 0x92, 0x98, 0x77, 0x32, 0x88, 0x3c, 0xe0, 0x1e, 0x63, 0x3c, 0x33, 0x23, 0x5c, 0xe5, 0xef, 0xa4, 0x69, 0x1e, 0x3d, 0x12, 0x62, 0xf9, 0x9f, 0x1a, 0xbd, 0xa4, 0xfa, 0x94, 0xb0, 0xf1, 0x79, 0x72, 0xb4, 0xf9, 0xc, 0xe3, 0x96, 0xfd, 0x10, 0x91, 0x4, 0xbc, 0x93, 0x66, 0x9, 0x8, 0x92, 0xb0, 0x98, 0x47, 0x8f, 0x84, 0x58, 0x54, 0xd7, 0xd1, 0x69, 0x54, 0x44, 0x43, 0xa1, 0x11, 0x93, 0xe1, 0xd3, 0xb8, 0xe5, 0x3f, 0x24, 0x24, 0x9d, 0x0, 0xde, 0x49, 0x7, 0xb6, 0x39, 0xa7, 0x87, 0x78, 0xfd, 0xc4, 0x24, 0xd1, 0xb4, 0x8b, 0x98, 0x0, 0x3a, 0xb7, 0xf3, 0xdc, 0x37, 0x97, 0x76, 0xff, 0xb8, 0x73, 0x29, 0x26, 0x9f, 0xdb, 0xf9, 0x8, 0x58, 0x38, 0x87, 0x9b, 0x1a, 0x34, 0x81, 0xbe, 0xc0, 0xce, 0x64, 0x96, 0x9e, 0x29, 0xfb, 0xfc, 0x36, 0x9d, 0x69, 0xd7, 0x3a, 0x73, 0x9d, 0xb7, 0x3c, 0xae, 0x79, 0xdc, 0xfd, 0xbf, 0xff, 0xd7, 0xcc, 0x48, 0x3e, 0xe0, 0xb8, 0x31, 0x30, 0x4, 0x2c, 0xf5, 0xbc, 0x93, 0xe, 0xd6, 0xda, 0x72, 0xcc, 0x1, 0xc9, 0x3c, 0x94, 0xe7, 0xd1, 0xb6, 0x8, 0xa6, 0xf4, 0xc1, 0x2f, 0xff, 0xba, 0x70, 0xa6, 0xfa, 0xda, 0x7c, 0x5c, 0x2e, 0xce, 0xbf, 0xb7, 0xf0, 0xe0, 0x97, 0xf2, 0x55, 0x67, 0xd5, 0x68, 0x42, 0x22, 0x89, 0xcc, 0xd2, 0x93, 0xe5, 0x7c, 0x96, 0xfe, 0x0, 0x4, 0x7f, 0xb5, 0x71, 0x72, 0xdd, 0xe9, 0x65, 0xd6, 0xc5, 0x60, 0x7, 0xd6, 0xd7, 0xdd, 0x19, 0x68, 0x12, 0x68, 0xba, 0x81, 0xe6, 0x4e, 0xca, 0xad, 0xaf, 0x88, 0x82, 0xae, 0x92, 0x1b, 0x51, 0xda, 0x47, 0xa2, 0x5f, 0x6d, 0x5d, 0xbe, 0x38, 0x55, 0x29, 0xce, 0x17, 0xab, 0x3e, 0xcf, 0x73, 0xaf, 0xaa, 0x21, 0xcb, 0x9f, 0xcd, 0x7, 0xec, 0x7f, 0x32, 0x55, 0x59, 0xbf, 0x2e, 0x5f, 0x95, 0x79, 0xc0, 0x1d, 0xfc, 0x79, 0x36, 0xc5, 0xa7, 0xc1, 0x22, 0x26, 0x67, 0xe9, 0xdd, 0x74, 0xa, 0x41, 0xd3, 0x8e, 0x3f, 0x7e, 0x3a, 0x24, 0x26, 0x6f, 0x1e, 0x75, 0xa7, 0xa1, 0xfe, 0x1b, 0x37, 0x12, 0x6, 0x8a, 0xa4, 0xc7, 0x9b, 0xdc, 0x5c, 0x3a, 0xa2, 0x9f, 0x8f, 0xcd, 0xa3, 0x9f, 0x8d, 0x1a, 0xbb, 0x92, 0xe8, 0x8f, 0x3b, 0xcd, 0xda, 0x54, 0xa5, 0xda, 0xc3, 0x47, 0xa0, 0x6f, 0x8a, 0xc1, 0x5f, 0xfc, 0x93, 0xfe, 0x11, 0xef, 0x7b, 0xd5, 0x6c, 0xe9, 0xcd, 0x6b, 0xf2, 0x55, 0xd7, 0x5a, 0x79, 0xb7, 0x80, 0xfb, 0x54, 0x85, 0xd7, 0x19, 0x77, 0x56, 0xb2, 0xf4, 0x16, 0xe8, 0x3d, 0xf4, 0x61, 0x7a, 0x8c, 0x3e, 0x4a, 0xbf, 0x0, 0x68, 0x9a, 0xe4, 0x35, 0x9e, 0x32, 0x3e, 0xd3, 0x3f, 0xa3, 0xa7, 0xc7, 0x9c, 0x40, 0x87, 0xdf, 0xed, 0x4e, 0xdc, 0x59, 0x68, 0x92, 0xda, 0xdc, 0xb7, 0xb9, 0xf9, 0x76, 0xb1, 0xcb, 0x7f, 0x4b, 0x92, 0x79, 0x28, 0xd1, 0x47, 0xed, 0xfb, 0xde, 0xce, 0x96, 0xaa, 0x1, 0x5a, 0xc4, 0x31, 0xc, 0x1f, 0xd9, 0xfa, 0x47, 0xfc, 0xd7, 0xd5, 0x5c, 0xe9, 0xe5, 0xab, 0xf2, 0x55, 0x1f, 0xd8, 0x2d, 0xa9, 0x56, 0x66, 0xd, 0x3e, 0x61, 0xb9, 0x31, 0x57, 0xa5, 0x71, 0x67, 0x9d, 0x1e, 0xf7, 0x90, 0xf4, 0xb0, 0x8f, 0xa6, 0x47, 0x28, 0x50, 0x78, 0xf4, 0x42, 0xcb, 0xc2, 0xb6, 0x57, 0xfa, 0xb1, 0x98, 0x45, 0x70, 0x68, 0xe3, 0xa, 0xb8, 0xea, 0xd8, 0x31, 0x31, 0x0, 0x2c, 0x85, 0xde, 0x49, 0x9d, 0x87, 0x5b, 0x99, 0xad, 0x84, 0xcc, 0xa3, 0x63, 0xdd, 0x5, 0xa5, 0xf5, 0xce, 0x63, 0x8b, 0xc, 0x4d, 0x18, 0x7a, 0x70, 0x3c, 0xc9, 0xe7, 0x4c, 0x55, 0x7e, 0xb1, 0x2a, 0xb, 0xe8, 0x44, 0x7b, 0xb1, 0xbf, 0x17, 0x2, 0xf7, 0xc, 0xbd, 0xd1, 0xf, 0x5a, 0x7a, 0x5f, 0x78, 0x18, 0xea, 0x61, 0xa9, 0xfb, 0xea, 0x73, 0x20, 0xf6, 0xf7, 0xe5, 0xf1, 0x53, 0x9d, 0x5d, 0x87, 0xb4, 0x63, 0xd9, 0xb7, 0xa, 0x5e, 0xed, 0xdc, 0x11, 0x68, 0xe2, 0xe8, 0x66, 0xe0, 0x9d, 0x74, 0x30, 0xed, 0x8f, 0xb5, 0x10, 0x97, 0x8b, 0x3b, 0x80, 0xcc, 0x43, 0x89, 0xce, 0x2e, 0x75, 0x75, 0x53, 0x91, 0xd3, 0x39, 0xb2, 0xa6, 0x52, 0xe1, 0x8a, 0x9d, 0x99, 0xab, 0x5c, 0xfb, 0x97, 0x7c, 0xcd, 0xb9, 0xe, 0x9b, 0x55, 0x20, 0xf1, 0xc5, 0x40, 0xa6, 0x62, 0xc4, 0xe4, 0xb6, 0x67, 0x25, 0x1d, 0xb, 0x91, 0xd4, 0x45, 0xd3, 0x71, 0xfa, 0x1, 0x10, 0x7b, 0xbb, 0x55, 0xe8, 0x6a, 0x3c, 0x6d, 0x4f, 0x53, 0xe3, 0x29, 0xef, 0x59, 0xfb, 0x88, 0x87, 0x7d, 0xdc, 0xc8, 0x48, 0x80, 0x24, 0x4a, 0x97, 0xb6, 0xc3, 0xbc, 0xe8, 0x4d, 0x1e, 0x27, 0x8a, 0x5e, 0x26, 0xcc, 0xe2, 0x6, 0xb2, 0xa3, 0x8c, 0x44, 0xff, 0x58, 0xce, 0x95, 0xf4, 0x1a, 0x49, 0xa7, 0xa3, 0xaa, 0xf3, 0xd9, 0xca, 0xd9, 0x25, 0xf9, 0x9a,
-    0x57, 0xbc, 0xfe, 0x92, 0x41, 0xf6, 0x69, 0xc1, 0x56, 0xc2, 0x7d, 0x41, 0x42, 0x52, 0xf, 0x4d, 0xcb, 0x40, 0xec, 0x47, 0xda, 0x25, 0xf5, 0xee, 0x2f, 0x16, 0xd8, 0x62, 0x15, 0x3a, 0xf, 0xec, 0x1e, 0x68, 0x34, 0x9, 0xc4, 0x65, 0xfb, 0x44, 0xdb, 0x57, 0xc2, 0x6f, 0x8b, 0xf2, 0x60, 0x76, 0x94, 0x91, 0xe8, 0xfe, 0x8f, 0x4f, 0x55, 0x6c, 0x2c, 0x26, 0xfc, 0x88, 0x87, 0xa6, 0xd2, 0xd3, 0x8b, 0x37, 0xa5, 0xfb, 0xaa, 0xb3, 0x5c, 0x4b, 0x57, 0xd6, 0x22, 0x30, 0x1f, 0xfd, 0x75, 0xdf, 0xe6, 0x86, 0x7c, 0xdc, 0x43, 0xd8, 0x5, 0x20, 0xf6, 0x99, 0x4e, 0x90, 0xe5, 0xa5, 0xf3, 0xa1, 0xf4, 0x5f, 0x21, 0x9e, 0x88, 0x3c, 0x16, 0xb3, 0x19, 0x37, 0x42, 0x62, 0xc9, 0xed, 0x44, 0x2f, 0xdb, 0x47, 0xdf, 0xe, 0x28, 0xfb, 0xeb, 0xa3, 0xed, 0xd, 0xa8, 0x15, 0x24, 0x7a, 0xf9, 0x6a, 0x36, 0x44, 0x93, 0xde, 0x52, 0xc2, 0x8e, 0x32, 0x34, 0xdd, 0xf7, 0x36, 0xf4, 0x5f, 0x4e, 0x37, 0xc4, 0x4c, 0x1a, 0xb8, 0x77, 0xe0, 0x7, 0x88, 0x4e, 0xea, 0xf3, 0x43, 0x1e, 0xd2, 0x64, 0x7a, 0xab, 0xb3, 0x45, 0xa0, 0xc6, 0x3, 0x6d, 0x8a, 0x21, 0x2b, 0xd0, 0x8c, 0x48, 0xb4, 0xe0, 0xa0, 0xa0, 0x89, 0xa3, 0x1b, 0x14, 0xed, 0x55, 0x5a, 0x2c, 0x39, 0xe1, 0xdf, 0x8c, 0x7b, 0x61, 0x90, 0x99, 0x87, 0x12, 0xbd, 0x79, 0x2d, 0x5b, 0x2a, 0x56, 0x31, 0xd4, 0xc8, 0xd6, 0x13, 0x86, 0xaf, 0xaa, 0x67, 0x37, 0x35, 0x6b, 0x97, 0xc0, 0xf8, 0xb1, 0xd4, 0xdb, 0x1, 0xc8, 0xb7, 0xf4, 0xe4, 0xbd, 0x3, 0xaf, 0xb, 0x36, 0x37, 0xc6, 0xf7, 0xd0, 0x17, 0x81, 0xd8, 0x57, 0xd8, 0xf8, 0xe9, 0xa, 0xb6, 0x3d, 0xac, 0x58, 0xd0, 0xf8, 0x25, 0x1c, 0x64, 0xa5, 0xb9, 0x3, 0x81, 0x26, 0x81, 0x80, 0x77, 0x12, 0x8f, 0xdc, 0x22, 0xec, 0xcf, 0xa3, 0x7, 0x5d, 0x63, 0x22, 0xd1, 0xfa, 0x9a, 0x37, 0xd2, 0x55, 0x21, 0x6e, 0xcc, 0x58, 0xea, 0xa2, 0x69, 0xaa, 0xf2, 0x9f, 0x8b, 0xcf, 0x7d, 0x23, 0x5f, 0xb5, 0xeb, 0x69, 0x2c, 0x23, 0x96, 0xde, 0xf, 0xf4, 0x51, 0x3, 0x92, 0xba, 0x68, 0x7a, 0x1, 0x88, 0x9d, 0xe5, 0x26, 0x4, 0x68, 0xa, 0xdb, 0xce, 0x41, 0x22, 0xdf, 0xaa, 0x1e, 0xeb, 0x60, 0x6b, 0x85, 0xef, 0x77, 0x34, 0x9, 0x34, 0xd9, 0xd4, 0x64, 0xfa, 0x48, 0xed, 0x20, 0x23, 0x29, 0x52, 0xc4, 0x24, 0x36, 0x9a, 0xbe, 0xab, 0x9f, 0xa9, 0x4e, 0x57, 0xab, 0x56, 0xfe, 0x0, 0x19, 0x4b, 0x3e, 0x9a, 0xaa, 0xaf, 0x21, 0xb1, 0x15, 0xb6, 0x83, 0x4b, 0x1e, 0xd9, 0x3b, 0xf0, 0x29, 0xed, 0xf8, 0xd6, 0xe7, 0x7b, 0xbd, 0x33, 0x6f, 0x3, 0xc1, 0x4b, 0xb9, 0x9, 0x4a, 0xf, 0xa6, 0xca, 0x26, 0xcd, 0xb8, 0x9f, 0x1e, 0xa4, 0xd8, 0x8a, 0x40, 0x9f, 0xa9, 0xbc, 0x6d, 0x72, 0xad, 0x18, 0xd2, 0x22, 0xde, 0x3c, 0xda, 0x5d, 0x19, 0xce, 0x53, 0x4b, 0xf4, 0xcb, 0xad, 0xcb, 0xb, 0x5d, 0x6f, 0x78, 0x1f, 0x23, 0x6a, 0x7b, 0x89, 0x3f, 0x5a, 0x9c, 0xef, 0x7a, 0xc3, 0xcf, 0x54, 0x6f, 0x1, 0x57, 0xe3, 0x6c, 0xeb, 0xe4, 0xba, 0x5c, 0x4d, 0x7c, 0x56, 0x61, 0x73, 0x63, 0x7c, 0x8c, 0x4e, 0x63, 0x91, 0x3a, 0x39, 0x37, 0xc1, 0x26, 0xf7, 0x2b, 0x7c, 0x87, 0x8d, 0xb9, 0xfb, 0x18, 0x4d, 0x1c, 0x31, 0xef, 0x24, 0x1a, 0x5b, 0xc0, 0x32, 0x28, 0xa5, 0xf3, 0x86, 0xbc, 0xa3, 0x8c, 0x44, 0xbf, 0xdd, 0xf9, 0xf4, 0x22, 0x43, 0x93, 0xce, 0xee, 0x2e, 0x22, 0x56, 0x78, 0x31, 0xb0, 0xc2, 0x59, 0x6c, 0xe5, 0xc3, 0x35, 0xf9, 0xaa, 0xb2, 0x4e, 0xfa, 0x67, 0x4, 0x24, 0x75, 0xd1, 0xf4, 0x24, 0x75, 0x81, 0xe0, 0x91, 0xdc, 0x4, 0x2c, 0xeb, 0x4b, 0xd1, 0xba, 0xcc, 0x6e, 0xc8, 0x1d, 0x8c, 0xd8, 0x8a, 0x40, 0x13, 0x1b, 0x8a, 0xd5, 0x94, 0xf0, 0xec, 0x43, 0x49, 0x1b, 0xbf, 0x3f, 0xdc, 0x6a, 0x1f, 0x89, 0x2e, 0xed, 0xde, 0xf7, 0x76, 0xae, 0x54, 0x95, 0x74, 0xe, 0x8f, 0x21, 0xe8, 0xcd, 0x14, 0x6c, 0x2b, 0x34, 0x52, 0xc7, 0xd3, 0xe7, 0x16, 0x76, 0x12, 0x44, 0xd3, 0x23, 0x48, 0xa4, 0xe, 0xe4, 0x26, 0xe0, 0xbe, 0x1, 0xe1, 0x93, 0xf0, 0x7c, 0x5f, 0x37, 0x4d, 0x6c, 0x20, 0xb9, 0xe6, 0xe3, 0xc6, 0x8e, 0x46, 0x3e, 0x57, 0x79, 0xef, 0x24, 0xd6, 0x8f, 0x94, 0x7a, 0xd9, 0xdf, 0x51, 0x66, 0xc8, 0x6b, 0x35, 0x9c, 0x7e, 0x41, 0x6e, 0xcd, 0x7a, 0xe7, 0xb1, 0x30, 0x52, 0x7, 0x35, 0x10, 0x66, 0x2b, 0x49, 0x9f, 0x54, 0xa7, 0x2a, 0xbf, 0xfe, 0x44, 0x85, 0x24, 0xd7, 0x43, 0x45, 0x74, 0x2c, 0xa9, 0x62, 0x2b, 0x8a, 0xdc, 0x4, 0x55, 0x7f, 0x95, 0x98, 0x59, 0xa2, 0x5b, 0x64, 0x5f, 0xe7, 0x38, 0x81, 0xa7, 0x2d, 0x41, 0xeb, 0x10, 0x7f, 0x7e, 0xa1, 0x17, 0x61, 0x63, 0xfa, 0x9c, 0x47, 0x83, 0xbc, 0xd7, 0xfa, 0xf9, 0x99, 0x57, 0x60, 0x46, 0xda, 0xcd, 0xce, 0xd3, 0xef, 0xf6, 0x63, 0x2b, 0x50, 0x13, 0x41, 0x2b, 0x4a, 0x1e, 0xf1, 0xb2, 0x95, 0x5d, 0xb4, 0xb8, 0x4, 0x46, 0x4c, 0xa2, 0xa0, 0xe9, 0x38, 0xfd, 0x10, 0x5c, 0x31, 0xcc, 0x4d, 0x8, 0xe7, 0xfc, 0xda, 0xfe, 0x9, 0x7a, 0x2b, 0xb3, 0x46, 0x67, 0xf6, 0x6b, 0x6c, 0x45, 0xa0, 0x67, 0x99, 0x77, 0x32, 0x62, 0x4e, 0x49, 0xaf, 0x45, 0xb0, 0xcc, 0xc3, 0x3a, 0x4d, 0xbe, 0x12, 0xae, 0xdd, 0x15, 0x5e, 0xff, 0x7b, 0xb6, 0x82, 0x65, 0x11, 0xd8, 0x69, 0x27, 0x16, 0x5b, 0xf9, 0x3b, 0xc, 0x83, 0xd0, 0xdf, 0xd3, 0xbb, 0x63, 0x22, 0xa9, 0x87, 0xa6, 0xf7, 0xc1, 0x35, 0x1f, 0xe8, 0xc7, 0x56, 0x62, 0xb4, 0x34, 0x6b, 0xe7, 0x12, 0xb6, 0x66, 0xd5, 0xb8, 0x71, 0x24, 0x61, 0xa9, 0x4e, 0xb7, 0x64, 0xef, 0xa4, 0x7e, 0xf, 0x53, 0xee, 0x4c, 0x16, 0x31, 0x79, 0xab, 0x63, 0x87, 0x8e, 0xe1, 0xd0, 0xee, 0x4a, 0xae, 0x2, 0x47, 0x30, 0x1d, 0x9a, 0xf8, 0xb3, 0x99, 0x37, 0xfc, 0xc5, 0x77, 0xe5, 0x6b, 0x7e, 0xef, 0xcd, 0xc9, 0xe2, 0xea, 0xa5, 0x2e, 0x9a, 0x8e, 0x79, 0x96, 0xbb, 0x4c, 0x61, 0x6e, 0x82, 0x6a, 0xe5, 0x2, 0x45, 0x8b, 0xf7, 0x7a, 0x39, 0x8b, 0xad, 0x7c, 0xb6, 0xff, 0xbc, 0xe1, 0x2, 0x75, 0x57, 0xe8, 0x76, 0x30, 0xef, 0xac, 0x6a, 0xa4, 0xb, 0xf5, 0x52, 0xc6, 0x7d, 0x60, 0x34, 0x75, 0x63, 0x1a, 0x7a, 0x62, 0x35, 0x57, 0xb1, 0xd1, 0x4c, 0xb8, 0x1d, 0xce, 0xd0, 0xf4, 0xd8, 0x62, 0x5d, 0xea, 0xf, 0xb7, 0xe9, 0xef, 0x12, 0xa1, 0x89, 0xd9, 0xe1, 0x6f, 0x80, 0x3b, 0x9d, 0xeb, 0xfc, 0x1b, 0xcb, 0x4d, 0x90, 0x7b, 0xaf, 0x72, 0x2f, 0x17, 0x16, 0x5b, 0x19, 0xe1, 0xda, 0x55, 0x91, 0x91, 0x74, 0x5a, 0xf6, 0x4e, 0xda, 0x6a, 0x60, 0x34, 0xf3, 0x70, 0x14, 0x4, 0x4d, 0x9c, 0x57, 0xfc, 0xd8, 0x8a, 0x39, 0xc7, 0x49, 0x85, 0xa6, 0x1c, 0x1a, 0xa9, 0x7b, 0x21, 0xa2, 0x4f, 0x40, 0xe6, 0x7b, 0xe8, 0x3c, 0xb8, 0x53, 0x55, 0x6e, 0x2, 0xd2, 0xfa, 0x7d, 0x9f, 0x1, 0x87, 0x33, 0xd6, 0x7b, 0x67, 0xf7, 0x4f, 0x6c, 0x45, 0xa0, 0xf6, 0x5e, 0xe4, 0xfd, 0x3, 0x83, 0x27, 0x45, 0x23, 0x26, 0x43, 0xa6, 0x5a, 0x67, 0xb9, 0x55, 0x6c, 0x2c, 0x42, 0xff, 0x3, 0xfd, 0x70, 0xad, 0x87, 0xa6, 0x68, 0x9a, 0xa9, 0x1a, 0xfc, 0x9d, 0xaa, 0x7c, 0x5b, 0x3b, 0xb7, 0x23, 0x5f, 0x35, 0x8a, 0xaf, 0x12, 0x47, 0x93, 0x36, 0xb6, 0xa2, 0xc2, 0x12, 0x9e, 0x1f, 0xc6, 0x8d, 0x5, 0x48, 0xbd, 0xde, 0xd8, 0xb1, 0xc4, 0x56, 0x27, 0x2e, 0xe8, 0xe2, 0x6d, 0xfd, 0x8c, 0x13, 0xf0, 0x7c, 0x79, 0xf2, 0xd4, 0x8, 0xea, 0xc6, 0x56, 0x3c, 0xfc, 0xcc, 0xb4, 0x6f, 0xed, 0xcd, 0xb6, 0x96, 0xb6, 0x27, 0x36, 0x4a, 0x9e, 0x8e, 0x3f, 0x49, 0xa, 0x68, 0x14, 0xfd, 0xd6, 0xf5, 0xa9, 0x6a, 0xb1, 0x5a, 0x95, 0xec, 0xeb, 0x62, 0xf8, 0xbf, 0x8c, 0x23, 0xd9, 0xa, 0x9f, 0xaa, 0x5c, 0x5e, 0xf8, 0xd5, 0x96, 0x7c, 0xd5, 0x37, 0xe8, 0xbd, 0x89, 0xd0, 0x74, 0xaf, 0x37, 0x56, 0xca, 0x74, 0x93, 0x4e, 0x6f, 0x4, 0xb1, 0x15, 0xd9, 0x7, 0xac, 0x8a, 0x2a, 0x88, 0x23, 0x1d, 0xb2, 0xe2, 0xa, 0x1d, 0x7, 0x9a, 0x38, 0xaa, 0xf7, 0xbd, 0x93, 0x98, 0x5f, 0x1b, 0x8b, 0xc0, 0x71, 0xcf, 0xe4, 0x90, 0xdc, 0xf6, 0x74, 0xe3, 0xe7, 0x43, 0xe2, 0x5c, 0x63, 0xa2, 0xb1, 0xb5, 0xb9, 0x45, 0xde, 0xf1, 0x7e, 0x39, 0xdf, 0xdd, 0x57, 0xc9, 0xb3, 0x34, 0x9c, 0xcd, 0x72, 0xd8, 0x9a, 0x30, 0x8a, 0xfe, 0xe0, 0x97, 0xaf, 0xcd, 0x4f, 0x87, 0xba, 0x9, 0xb3, 0xc7, 0xf5, 0x78, 0x62, 0xb5, 0x2b, 0xdf, 0x81, 0xe9, 0xe3, 0x82, 0x67, 0xf9, 0xc4, 0xf1, 0x34, 0xf5, 0xd1, 0x34, 0x8d, 0x44, 0xea, 0x96, 0x1a, 0x5c, 0x86, 0x18, 0xd6,
-    0x77, 0x31, 0x4c, 0x9, 0x23, 0x1d, 0x52, 0x1, 0x3a, 0x36, 0x24, 0xb1, 0xdc, 0xc9, 0xbb, 0x6c, 0xf7, 0xf, 0x44, 0x7b, 0x49, 0x9e, 0xc, 0x9b, 0xb, 0x7e, 0xbd, 0x8f, 0xd3, 0xdb, 0x3f, 0x8e, 0x6b, 0x61, 0xbc, 0x6f, 0x3e, 0xf7, 0xcd, 0xe5, 0x20, 0xb6, 0x52, 0xb4, 0xc0, 0xe, 0xb4, 0x9b, 0x8a, 0xd5, 0x53, 0x48, 0x15, 0xd4, 0xfb, 0xfe, 0x2c, 0x3f, 0x3e, 0x9a, 0x8e, 0xd1, 0x53, 0xde, 0xcc, 0x50, 0xa6, 0x9f, 0x7e, 0x1d, 0xa1, 0xee, 0x13, 0x68, 0xaa, 0xf2, 0xb8, 0xab, 0xa0, 0x4, 0xfa, 0x43, 0x4b, 0xea, 0x19, 0xf8, 0xe8, 0x86, 0xf5, 0xc, 0x78, 0x26, 0xcc, 0xcc, 0x91, 0x57, 0xa8, 0x90, 0xdf, 0xc1, 0x91, 0x55, 0x17, 0x41, 0x47, 0xa3, 0x38, 0xc8, 0xe, 0x80, 0xf4, 0xd2, 0x6e, 0xf3, 0x2f, 0xb9, 0x8a, 0x3a, 0x8b, 0x40, 0x87, 0xa6, 0x5e, 0xa4, 0xee, 0x3, 0x10, 0xa9, 0xfb, 0x30, 0x31, 0x9a, 0xb0, 0x2a, 0xa8, 0x67, 0xb0, 0xba, 0x4f, 0xdc, 0xe2, 0x40, 0xfe, 0x67, 0x55, 0x50, 0xd9, 0x71, 0xa1, 0x49, 0xa0, 0x23, 0x72, 0xee, 0xa4, 0xcd, 0x9a, 0x5c, 0xea, 0xcf, 0xd5, 0x58, 0x30, 0x5f, 0x31, 0xf2, 0xd9, 0x8e, 0xc6, 0xe3, 0x4e, 0xe9, 0x47, 0x9d, 0x6e, 0x15, 0x94, 0xde, 0xa, 0xd7, 0xa1, 0xe9, 0x54, 0x5, 0x56, 0x41, 0x7d, 0x4e, 0x4d, 0x19, 0x71, 0x26, 0x34, 0x61, 0x55, 0x50, 0xb3, 0x30, 0x2f, 0x83, 0x8, 0xcf, 0xa9, 0x6d, 0xe7, 0x6e, 0x15, 0x54, 0xd, 0x5c, 0x75, 0xb4, 0x58, 0xe2, 0xd6, 0x9d, 0xc4, 0xfb, 0xbf, 0x72, 0xa4, 0x46, 0x11, 0xa7, 0x5e, 0xfb, 0x51, 0x1d, 0x8f, 0x91, 0x63, 0x9, 0x86, 0x98, 0x9f, 0xd0, 0x8e, 0xe1, 0xbd, 0x30, 0x34, 0xc1, 0x28, 0x7a, 0xbf, 0xa, 0xca, 0xc4, 0xb8, 0x45, 0xce, 0x62, 0x2b, 0xf7, 0x7f, 0x2c, 0xb, 0xe8, 0xb, 0xf, 0xd, 0xc9, 0xfc, 0x97, 0xf, 0xd3, 0x75, 0x20, 0x76, 0x50, 0x5, 0x85, 0xdb, 0x4f, 0xb0, 0xd, 0x2, 0x34, 0x8d, 0xa5, 0xa, 0x4a, 0xa0, 0x5c, 0x43, 0xeb, 0xe3, 0xc0, 0x10, 0xa6, 0x1b, 0x85, 0x54, 0xb3, 0xf, 0x55, 0xef, 0x32, 0xfd, 0x9a, 0xe9, 0x5e, 0xb8, 0x4f, 0xbb, 0x7d, 0x13, 0x7a, 0x9c, 0xce, 0x2e, 0xe5, 0x50, 0x34, 0x61, 0x63, 0x5c, 0x71, 0xbe, 0xe7, 0x1b, 0xe8, 0x7f, 0x92, 0x45, 0xaa, 0xa0, 0x5c, 0xfa, 0x64, 0x42, 0x34, 0x19, 0xab, 0xa0, 0xf0, 0x1e, 0xa4, 0x69, 0xa7, 0xf2, 0xe8, 0xab, 0xa0, 0x4, 0xa, 0x76, 0xc5, 0x51, 0x69, 0x20, 0xf5, 0xcc, 0xc2, 0x14, 0x57, 0xc1, 0x75, 0x97, 0x1d, 0x62, 0xf4, 0xa3, 0x24, 0xde, 0x57, 0x9, 0x9b, 0x4d, 0xaa, 0xa2, 0xe8, 0xd7, 0xfe, 0x95, 0xad, 0xe0, 0xfe, 0x1, 0x9b, 0x9c, 0x2, 0xe6, 0xd, 0x87, 0x55, 0x50, 0x3f, 0x24, 0x88, 0xf9, 0xf6, 0xd0, 0xa4, 0xad, 0x82, 0xd2, 0x69, 0x28, 0x65, 0x8b, 0xe6, 0xf1, 0x7d, 0x8a, 0x46, 0x80, 0xa4, 0x70, 0xdd, 0x49, 0x1b, 0x99, 0xe2, 0x96, 0x33, 0xca, 0x8e, 0x8d, 0xcd, 0xa4, 0x46, 0x89, 0x9, 0x7f, 0x61, 0xe5, 0x7d, 0xa9, 0xdf, 0xb, 0x82, 0xef, 0x74, 0xfb, 0x26, 0x58, 0xbb, 0x80, 0xde, 0xff, 0x71, 0xb6, 0x54, 0x44, 0x91, 0xa3, 0xc7, 0x92, 0x6f, 0x35, 0x5, 0x55, 0x50, 0x30, 0xb6, 0xf2, 0x7c, 0x62, 0x8f, 0x93, 0xaa, 0xa, 0xaa, 0xac, 0x6a, 0xd, 0xb5, 0xed, 0x1a, 0xb4, 0xf9, 0x8, 0xab, 0xa0, 0x84, 0xbb, 0x2e, 0xf4, 0xc6, 0x67, 0x83, 0x47, 0x2c, 0x26, 0x9b, 0xa2, 0xc2, 0x3c, 0x2e, 0x45, 0xf, 0x96, 0xea, 0x9e, 0xf4, 0xf3, 0xbb, 0x10, 0x4d, 0x25, 0xa4, 0x42, 0xf1, 0x37, 0x9f, 0x74, 0xeb, 0x56, 0x54, 0x23, 0x9b, 0xfa, 0x7d, 0x4f, 0x37, 0xd, 0x27, 0xb6, 0x2, 0xab, 0xa0, 0x6a, 0x9d, 0x43, 0xd8, 0x6e, 0x7c, 0xc4, 0xd0, 0x6, 0x7d, 0x8f, 0xd3, 0x68, 0xaa, 0xa0, 0x38, 0x12, 0x76, 0xc5, 0x31, 0xfa, 0x30, 0xcc, 0x32, 0x14, 0xdf, 0x39, 0x76, 0x67, 0xc6, 0x41, 0x2c, 0xbc, 0x16, 0xb8, 0x8a, 0x83, 0xf6, 0xcd, 0x37, 0xaf, 0x61, 0xb1, 0x15, 0x9d, 0x5e, 0xe2, 0x47, 0x42, 0x55, 0x15, 0xd4, 0x8b, 0x89, 0xd1, 0x4, 0xab, 0xa0, 0x6e, 0x50, 0x64, 0x85, 0xd6, 0x8, 0xb3, 0xe9, 0x11, 0x54, 0x41, 0x9, 0x54, 0x6d, 0x8, 0xb9, 0x93, 0x96, 0xe3, 0x71, 0x49, 0xa7, 0x6d, 0x78, 0xab, 0x9c, 0xe8, 0xce, 0x76, 0x54, 0x18, 0x96, 0x5a, 0x4, 0xb9, 0x2b, 0xcc, 0xcb, 0x22, 0xce, 0xa, 0x82, 0x5f, 0xc0, 0xa2, 0xe8, 0xeb, 0xd7, 0xbb, 0x55, 0x50, 0x36, 0x58, 0x82, 0xf3, 0x3a, 0x16, 0x5b, 0xf9, 0x14, 0xa9, 0x82, 0x52, 0x55, 0xf2, 0xda, 0xa3, 0x29, 0x52, 0xa4, 0xce, 0xca, 0x2, 0x1d, 0x72, 0x15, 0x94, 0x40, 0x2f, 0xe9, 0xf6, 0xec, 0x8e, 0xef, 0x1d, 0x42, 0xe6, 0xaa, 0xa6, 0xd9, 0x7, 0xff, 0xda, 0x31, 0xfd, 0x96, 0xc9, 0x86, 0xe0, 0x2d, 0x71, 0xaf, 0x35, 0x5f, 0x2, 0x7d, 0x93, 0x55, 0x41, 0xf1, 0xba, 0x9, 0xf3, 0x4, 0xe0, 0x36, 0x7a, 0x80, 0x26, 0xf, 0x89, 0xb0, 0xa, 0xea, 0x42, 0x62, 0xbb, 0xe9, 0x79, 0x75, 0x15, 0x94, 0x8d, 0xee, 0x46, 0xec, 0x1, 0x7, 0xf5, 0xdf, 0xe, 0x1e, 0x49, 0xb7, 0xe2, 0x66, 0x4, 0xf0, 0x88, 0x51, 0x8d, 0x8b, 0x44, 0xde, 0xc5, 0x12, 0x39, 0x5b, 0x3f, 0x7, 0x34, 0xb4, 0x92, 0xf1, 0x4a, 0x5c, 0xdf, 0x94, 0x6b, 0xb8, 0xfd, 0x15, 0xe6, 0x84, 0x2a, 0x28, 0x3d, 0x43, 0xdd, 0xe4, 0x57, 0x41, 0x1, 0x17, 0xf3, 0x72, 0xc2, 0x48, 0xdd, 0x31, 0x74, 0x85, 0xb9, 0xe6, 0x36, 0xd7, 0xe3, 0xf5, 0xd6, 0x24, 0xe8, 0x65, 0x8e, 0x1f, 0x5b, 0x19, 0x4a, 0x15, 0x14, 0x47, 0xb5, 0xce, 0xa2, 0xaa, 0x5e, 0xc9, 0x46, 0xd6, 0xe2, 0x39, 0xa2, 0xf5, 0xdc, 0xff, 0xc4, 0xe6, 0xbb, 0x3a, 0xbc, 0xa9, 0xb3, 0x12, 0x90, 0x11, 0x4d, 0x7d, 0x7d, 0x64, 0xbd, 0x55, 0x7a, 0x6e, 0xe7, 0xdb, 0x9a, 0xaa, 0xa, 0xca, 0x6, 0x53, 0x78, 0x15, 0xd4, 0x7, 0x3, 0x88, 0xd4, 0x29, 0xab, 0xa0, 0xb0, 0x36, 0xb5, 0xb0, 0x71, 0x87, 0x52, 0x5, 0x25, 0x50, 0x76, 0x3, 0xec, 0x44, 0x89, 0x8d, 0x39, 0x96, 0x99, 0xb9, 0x8, 0x82, 0x6c, 0xd6, 0x1a, 0x52, 0x5b, 0x4c, 0x72, 0x45, 0x34, 0x66, 0x21, 0xc9, 0xf8, 0xd5, 0xfc, 0x6, 0x16, 0x45, 0xff, 0x73, 0xbb, 0x5b, 0x5, 0x65, 0xab, 0x8d, 0xa0, 0x6e, 0xc2, 0xaa, 0xa0, 0xd6, 0xe9, 0x30, 0x62, 0x2b, 0x2f, 0xa1, 0x2b, 0xb4, 0x4a, 0x9a, 0x5f, 0x95, 0xe9, 0xeb, 0xc7, 0x96, 0x10, 0xff, 0xed, 0x80, 0x90, 0xb4, 0xb3, 0x2d, 0xad, 0x93, 0xa0, 0x7a, 0x1d, 0xd, 0x4b, 0xe2, 0x67, 0xf1, 0xe6, 0x85, 0xbd, 0x5f, 0xd5, 0xdb, 0xf6, 0x9b, 0x9c, 0x55, 0x6e, 0x61, 0xc3, 0x97, 0x37, 0xef, 0x72, 0xb3, 0x1b, 0x2b, 0x1d, 0xb1, 0x15, 0xea, 0x9d, 0xfb, 0x84, 0x2a, 0x28, 0x1d, 0x86, 0x8a, 0xc8, 0x51, 0xb6, 0xc2, 0xdc, 0x13, 0x60, 0x85, 0xb9, 0xeb, 0x31, 0xab, 0x9f, 0x7a, 0x8c, 0x57, 0x41, 0x2d, 0x7, 0xb1, 0x15, 0x87, 0x60, 0x6d, 0x61, 0x90, 0x87, 0xd2, 0x7f, 0x9b, 0x1c, 0x49, 0x60, 0xef, 0x34, 0x35, 0x96, 0xec, 0x91, 0x32, 0x58, 0x8e, 0xe6, 0x8f, 0xd0, 0x62, 0xaa, 0xcc, 0x6c, 0x26, 0x52, 0x70, 0xaf, 0x36, 0x64, 0x2c, 0xf1, 0x55, 0x50, 0xb6, 0xda, 0x49, 0xc6, 0x5d, 0xb6, 0x4, 0xab, 0xa0, 0xbe, 0xa0, 0x8f, 0xe, 0x21, 0x52, 0x87, 0xae, 0xd0, 0x6a, 0x9a, 0xd3, 0x70, 0x3c, 0xa0, 0x2a, 0x28, 0x81, 0x4e, 0xa8, 0xea, 0x95, 0x70, 0x7b, 0xe, 0xbf, 0xc7, 0xf8, 0xcf, 0x14, 0xc7, 0x46, 0xd3, 0xff, 0x9a, 0x26, 0x1b, 0xa6, 0xec, 0xfb, 0x5, 0xa, 0xee, 0xd2, 0x36, 0xd2, 0x27, 0x3, 0x62, 0x55, 0x50, 0x2a, 0xa4, 0xc8, 0x38, 0xc2, 0x30, 0x87, 0x55, 0x41, 0x7d, 0xef, 0xd9, 0x3d, 0x49, 0x63, 0x2b, 0xb0, 0xa, 0x2a, 0x8c, 0xad, 0x74, 0x3d, 0xfe, 0x36, 0x73, 0xe5, 0x7e, 0x5b, 0x90, 0x1, 0x55, 0x41, 0x71, 0x74, 0x3, 0x66, 0x4, 0xf4, 0x7f, 0x2d, 0x8e, 0xf4, 0xb9, 0x73, 0x1c, 0xf1, 0xe9, 0xf0, 0x6b, 0xea, 0xe7, 0x64, 0x36, 0x77, 0x60, 0x6a, 0xc3, 0x1e, 0x8a, 0xbc, 0xb6, 0xcb, 0xb8, 0x8b, 0x64, 0x47, 0x83, 0xa3, 0x2e, 0xed, 0xae, 0x64, 0xb9, 0x15, 0xe6, 0xd4, 0xf5, 0x50, 0xb8, 0xef, 0xc0, 0x8f, 0xad, 0xbc, 0x7b, 0x53, 0xfa, 0x8d, 0xdb, 0x43, 0xa9, 0x82, 0x7a, 0xb6, 0xbd, 0xd8, 0x8f, 0xad, 0x44, 0x5b, 0xfd, 0xd2, 0xe7, 0xc4, 0xb1, 0x15, 0x81, 0x90, 0xb5, 0xde, 0xb5, 0x77, 0xe4, 0xc0, 0x7a, 0x1a, 0x35, 0x1a, 0xd4, 0xe3, 0xb7, 0xfa, 0xb9, 0xf5, 0x38, 0x51, 0xe0, 0xcd, 0xc1, 0x31, 0xe8, 0x63, 0x88, 0xb5, 0xd8, 0x49, 0xf, 0x45, 0x25, 0x52, 0x6c, 0x2c, 0xb7, 0xec, 0xf6, 0x78, 0x7b, 0xe2, 0x93, 0x29, 0x6e, 0x85, 0x39, 0xc,
-    0x4b, 0xea, 0xfc, 0xf0, 0xa0, 0xa, 0xea, 0x6d, 0xac, 0xa, 0x2a, 0xa9, 0xc7, 0x9, 0xab, 0x82, 0xda, 0xf2, 0x9e, 0x4e, 0xeb, 0xbf, 0x34, 0xc8, 0x33, 0x41, 0x15, 0x94, 0x40, 0x93, 0xcd, 0xa3, 0xd8, 0xfe, 0x81, 0x6a, 0x39, 0x87, 0x9a, 0xc5, 0xe9, 0xdf, 0x8d, 0xec, 0x9, 0x50, 0x56, 0x6f, 0x69, 0xd9, 0x14, 0x19, 0xd6, 0x20, 0x49, 0xe4, 0xb2, 0xcf, 0x6c, 0x8d, 0x7f, 0x4f, 0xf, 0x79, 0x2d, 0x95, 0x77, 0x57, 0x49, 0xd5, 0x43, 0x11, 0xd8, 0x55, 0x2c, 0x24, 0x6c, 0xd7, 0x43, 0x39, 0xb6, 0x52, 0x4, 0x7f, 0x7b, 0xb9, 0x4c, 0x32, 0xeb, 0xaa, 0xa0, 0x92, 0xc7, 0x56, 0x60, 0x15, 0x54, 0xb8, 0xc2, 0x5c, 0xaf, 0xdf, 0x21, 0xfd, 0x57, 0xb4, 0x61, 0x1c, 0xe9, 0x18, 0x5a, 0x79, 0x16, 0x11, 0x49, 0x6b, 0x2d, 0xb0, 0xaf, 0xc2, 0x66, 0x82, 0x7d, 0x16, 0xc6, 0xcc, 0xdd, 0xc, 0xf0, 0x8c, 0x8f, 0x1f, 0x86, 0xa0, 0x45, 0x32, 0xd1, 0x58, 0xda, 0x5e, 0x68, 0x9d, 0x68, 0x1b, 0xaa, 0xc5, 0xd1, 0xb6, 0xe9, 0x55, 0x41, 0xc5, 0x66, 0xf, 0x4d, 0x9f, 0x5e, 0xfc, 0x2d, 0x88, 0xad, 0xb0, 0x2a, 0xa8, 0xe3, 0x9, 0x58, 0x15, 0x5b, 0xb9, 0x2b, 0xc9, 0xe, 0x19, 0xe4, 0x28, 0x5c, 0x3b, 0xc4, 0x8c, 0x26, 0x8e, 0x5e, 0x6d, 0x2f, 0x26, 0xda, 0xef, 0x65, 0x7f, 0x70, 0x99, 0x2c, 0x92, 0x77, 0x3c, 0x5e, 0x25, 0x5f, 0x6d, 0x4e, 0x34, 0xa6, 0x1b, 0x7f, 0xda, 0x3e, 0xdc, 0x5c, 0x68, 0xdd, 0xda, 0x9b, 0xe9, 0x58, 0x8d, 0x66, 0x9a, 0x9e, 0xc6, 0x62, 0x2b, 0xef, 0x2d, 0xbc, 0xb6, 0x70, 0x39, 0x36, 0x9f, 0x99, 0xff, 0xcf, 0x45, 0xac, 0xa, 0xea, 0x61, 0xfa, 0x48, 0x2, 0x3e, 0x8e, 0xc4, 0x7d, 0x59, 0x6c, 0xa5, 0xec, 0xf5, 0x9e, 0xb8, 0x5c, 0x26, 0x85, 0xf5, 0xc3, 0xff, 0x8b, 0x84, 0x26, 0x81, 0x56, 0xbc, 0xb1, 0xb6, 0x76, 0xf0, 0xd9, 0x7b, 0xe, 0xc6, 0x31, 0xd6, 0xba, 0x30, 0xb6, 0x10, 0xdb, 0x21, 0xe3, 0xd2, 0x6e, 0x7c, 0x5e, 0x69, 0x9f, 0xdb, 0x81, 0x39, 0x4, 0x2c, 0x63, 0x2e, 0x19, 0x43, 0x4f, 0x38, 0xa3, 0x5a, 0xe7, 0xa, 0x4d, 0xc2, 0xe8, 0x8c, 0xc4, 0xa, 0x49, 0x3f, 0x72, 0xb2, 0xec, 0x6f, 0x29, 0xa5, 0x6d, 0xa4, 0xa1, 0x88, 0x36, 0x65, 0x4a, 0x69, 0xfb, 0x20, 0x64, 0x46, 0x51, 0xda, 0x5e, 0x2a, 0x4a, 0xdb, 0x86, 0x51, 0x14, 0x4, 0xa5, 0x2d, 0xa6, 0x26, 0xaf, 0x4d, 0xe6, 0xe6, 0x7e, 0x84, 0xed, 0x12, 0x17, 0x3f, 0x29, 0x9a, 0x74, 0x74, 0xfe, 0xff, 0x79, 0xf6, 0x4a, 0xb7
-};
-
-// all the following images have sizes of 48672
-const uint8_t center_n[1075] PROGMEM = {
-0x78, 0x9c, 0xed, 0x98, 0xc1, 0x6e, 0xdb, 0x40, 0x14, 0xc4, 0xfa, 0xe7, 0xfa, 0xf4, 0x16, 0x30, 0x16, 0x42, 0xdb, 0xd8, 0xb1, 0xc5, 0x99, 0xa5, 0xa4, 0x76, 0x79, 0xcf, 0xe, 0x1f, 0x21, 0x5f, 0xf2, 0xe3, 0xc7, 0x2b, 0x7e, 0x52, 0xb6, 0x97, 0x4f, 0x9e, 0x8, 0x3b, 0xc5, 0x5b, 0xba, 0x6a, 0xb9, 0xd5, 0x8d, 0x71, 0x8b, 0x6e, 0x17, 0x2c, 0x77, 0x93, 0x6e, 0x97, 0x2b, 0xb7, 0xba, 0x31, 0x6e, 0xd3, 0xed, 0x62, 0xe5, 0x6e, 0xd4, 0xed, 0x52, 0xe5, 0x56, 0x37, 0xc6, 0xad, 0xba, 0x5d,
-0xa8, 0xdc, 0xcd, 0xba, 0xa1, 0xe5, 0x19, 0xdc, 0xad, 0xdb, 0x55, 0xbe, 0xb8, 0xdb, 0x75, 0xbb, 0x48, 0xb9, 0x1b, 0x76, 0xbb, 0xc4, 0x6f, 0xf5, 0x8e, 0xdd, 0xae, 0xf0, 0xc5, 0xdd, 0xb2, 0xdb, 0x5, 0xca, 0xdd, 0xb4, 0xdb, 0xe9, 0xe5, 0x98, 0xb4, 0xb7, 0xf4, 0x92, 0xd5, 0x8d, 0x72, 0x6a, 0x39, 0xa6, 0xec, 0x2d, 0x7d, 0xcb, 0x89, 0xe5, 0x98, 0xb0, 0xb7, 0xd4, 0xf7, 0xe8, 0xe0, 0xf9, 0x76, 0x8b, 0xd, 0x4e, 0xfb, 0xe2, 0x98, 0xae, 0xb7, 0xf4, 0x96, 0x93, 0xca, 0x31, 0x59, 0x6f,
-0x69, 0x8e, 0x4b, 0x8e, 0xe7, 0xda, 0xad, 0xf5, 0x1b, 0xa7, 0x7c, 0x71, 0x4c, 0xd5, 0x5b, 0xfa, 0x88, 0x13, 0xca, 0x31, 0x51, 0x6f, 0x69, 0x9e, 0x4f, 0x86, 0xe7, 0xd9, 0x2d, 0xf5, 0x17, 0xfa, 0x17, 0xc7, 0x34, 0xbd, 0xa5, 0x8f, 0x91, 0xcb, 0x31, 0x49, 0x6f, 0x69, 0xae, 0x13, 0xc7, 0x73, 0xec, 0x56, 0x7a, 0x82, 0xfa, 0xc5, 0x31, 0x45, 0x6f, 0xe9, 0x10, 0x62, 0x39, 0x26, 0xe8, 0x2d, 0xcd, 0xf7, 0x62, 0x78, 0x7e, 0xdd, 0x42, 0x2f, 0xd0, 0xbe, 0x38, 0xa6, 0xe7, 0x2d, 0x1d, 0x46,
-0x2a, 0xc7, 0xe4, 0xbc, 0x25, 0xc7, 0xcd, 0xba, 0xc6, 0x5b, 0x2, 0x28, 0x5f, 0x1c, 0x53, 0xf3, 0x96, 0x10, 0x42, 0x39, 0x26, 0xe6, 0x2d, 0x41, 0xa6, 0x97, 0x63, 0x5a, 0xde, 0x12, 0x64, 0x75, 0xa3, 0x4c, 0x2e, 0xc7, 0xa4, 0xbc, 0xa5, 0x80, 0xa9, 0xe5, 0x98, 0x92, 0xb7, 0x14, 0xb0, 0xba, 0x51, 0x26, 0x96, 0x63, 0x42, 0xd6, 0x12, 0x37, 0x1c, 0x4c, 0x2b, 0xc7, 0x74, 0xac, 0xa5, 0xc4, 0xf1, 0xc1, 0xea, 0x46, 0x99, 0x54, 0x8e, 0xc9, 0x58, 0x4b, 0x99, 0xe5, 0x60, 0x4a, 0x39, 0xa6,
-0x62, 0x2d, 0xa5, 0x9e, 0xf, 0x56, 0x37, 0xca, 0x84, 0x72, 0x4c, 0xc4, 0x5a, 0xca, 0x4d, 0xb9, 0x6f, 0xff, 0x9a, 0x73, 0xba, 0x5d, 0xeb, 0xb7, 0xca, 0x34, 0xac, 0xa5, 0x86, 0xeb, 0xa0, 0x5c, 0x8e, 0x49, 0x58, 0x4b, 0x1d, 0x5b, 0xee, 0xdc, 0xbd, 0xe6, 0xbc, 0x6e, 0xd7, 0xf9, 0xad, 0x32, 0x5, 0x6b, 0xa9, 0xe5, 0x3b, 0x28, 0x96, 0x63, 0x2, 0xd6, 0x52, 0xcf, 0x78, 0x50, 0x2b, 0xc7, 0xe6, 0xad, 0xa5, 0xa6, 0xf3, 0x83, 0xd5, 0x8d, 0x52, 0x2a, 0xc7, 0xc6, 0xad, 0xa5, 0xae, 0xf5,
-0xa0, 0x52, 0x8e, 0x4d, 0x5b, 0x4b, 0x6d, 0x6f, 0x6e, 0xdf, 0xd9, 0xb7, 0x96, 0xda, 0xde, 0x83, 0xc2, 0x17, 0xc7, 0x86, 0xad, 0xa5, 0xbe, 0xf9, 0x20, 0x2e, 0xc7, 0x66, 0xad, 0xa5, 0x19, 0xee, 0xfc, 0x82, 0x7c, 0xdb, 0x5a, 0x9a, 0xe1, 0x3e, 0x8, 0xbf, 0x38, 0x36, 0x6a, 0x2d, 0xcd, 0xb1, 0x1f, 0x44, 0xe5, 0xd8, 0xa4, 0xb5, 0x34, 0xcb, 0x9f, 0x5f, 0x91, 0xed, 0x5a, 0x4b, 0xb3, 0xfc, 0x7, 0xc1, 0x17, 0xc7, 0x6, 0xad, 0xa5, 0x79, 0x17, 0xc, 0x70, 0x39, 0x36, 0x67, 0x2d, 0xcd,
-0xbc, 0x81, 0x5f, 0xc2, 0x37, 0xad, 0xa5, 0x99, 0x37, 0xc, 0xe0, 0x17, 0xc7, 0xc6, 0xac, 0xa5, 0xb9, 0x57, 0xc, 0x50, 0x39, 0x36, 0x65, 0x2d, 0xcd, 0xbe, 0xc3, 0xbd, 0xe6, 0x5f, 0xeb, 0x46, 0xbe, 0x38, 0x36, 0x74, 0x7c, 0x67, 0x76, 0x37, 0xbb, 0x1c, 0x9b, 0x39, 0xba, 0x42, 0x97, 0x8c, 0x5b, 0xd8, 0x45, 0xce, 0xa, 0x5d, 0x32, 0x6e, 0x19, 0x1c, 0xfc, 0xe2, 0xd8, 0xc8, 0xb1, 0xd, 0xbe, 0x64, 0x6c, 0xec, 0x1c, 0x2a, 0xc7, 0x26, 0x8e, 0xde, 0x43, 0x97, 0x9c, 0x95, 0x9d, 0x3,
-0xe5, 0xd8, 0xc0, 0xf1, 0x7b, 0x56, 0x37, 0x76, 0x8f, 0xd5, 0xcd, 0x2a, 0xc7, 0x9e, 0xb7, 0xee, 0x21, 0x3b, 0x4e, 0x39, 0xf6, 0xb8, 0x75, 0xd, 0xd9, 0x59, 0xdd, 0xec, 0xff, 0x56, 0xc, 0x3e, 0x2a, 0xc7, 0x9e, 0xb6, 0x6e, 0x21, 0x3b, 0xc9, 0x5d, 0x83, 0xf, 0xca, 0xb1, 0x87, 0xad, 0x4b, 0xc8, 0x4e, 0x76, 0xd9, 0x83, 0xd5, 0x8d, 0xf2, 0xb6, 0x1c, 0x7b, 0xd6, 0xba, 0x83, 0xec, 0xa4, 0xb7, 0x7d, 0xb6, 0x3c, 0xe7, 0xd5, 0xd6, 0x12, 0xd9, 0xc9, 0xaf, 0x7b, 0xf0, 0xe6, 0x8b, 0x63,
-0x8f, 0x5a, 0x37, 0x90, 0x9d, 0xfc, 0xba, 0xc1, 0xb7, 0xe5, 0xd8, 0x93, 0xd6, 0xd, 0xac, 0x56, 0x7e, 0xdf, 0xfb, 0xf5, 0xfe, 0x8b, 0xcd, 0x25, 0xd6, 0xaa, 0x71, 0xe1, 0x83, 0x6f, 0xbe, 0x38, 0xf6, 0xa0, 0xe5, 0x4f, 0x5b, 0xe5, 0x17, 0xe, 0x5e, 0x96, 0x63, 0xcf, 0x59, 0xfe, 0xbc, 0x56, 0x7e, 0xe3, 0xe0, 0x45, 0x39, 0xf6, 0x98, 0x65, 0x9f, 0xd4, 0xca, 0xaf, 0x7c, 0xb0, 0xba, 0x51, 0x9e, 0x96, 0x63, 0x4f, 0x59, 0xee, 0x59, 0xad, 0xfc, 0xce, 0xc1, 0x93, 0x72, 0xec, 0x21, 0xcb,
-0x3c, 0xad, 0x95, 0x5f, 0xfa, 0xca, 0xa3, 0xf5, 0xce, 0x1c, 0xef, 0xb4, 0x55, 0x7e, 0xe9, 0xe0, 0xcb, 0x17, 0xc7, 0x9e, 0xb1, 0xbc, 0xf3, 0x5a, 0xf9, 0xad, 0x83, 0xad, 0xf1, 0x96, 0x65, 0xdd, 0xa8, 0x95, 0x5f, 0xfb, 0xcc, 0xa5, 0xf1, 0xc6, 0x3c, 0xe7, 0x46, 0xab, 0xfc, 0xda, 0xc1, 0x96, 0xbf, 0x64, 0x39, 0x77, 0x6a, 0xa5, 0x16, 0x3b, 0x5b, 0xfa, 0x8e, 0x65, 0xdc, 0xaa, 0x95, 0x7a, 0x7c, 0xf5, 0x49, 0xff, 0x7e, 0xae, 0x6f, 0xab, 0x55, 0xea, 0xb1, 0xb3, 0x65, 0xaf, 0x58, 0xbe,
-0xbd, 0x5a, 0xa9, 0xc9, 0xce, 0x96, 0xbc, 0x61, 0xd9, 0x36, 0x6b, 0xa5, 0x2e, 0x7f, 0x3a, 0x25, 0x7f, 0x3b, 0xdf, 0xb5, 0xd9, 0x2a, 0x75, 0xd9, 0xd9, 0xf8, 0xb, 0x96, 0x6b, 0xb7, 0x56, 0x6a, 0xb3, 0xb3, 0xfd, 0xbf, 0xdd, 0xd2, 0xdf, 0x2a, 0xfd, 0x3b, 0xc7, 0xb3, 0xdd, 0x2a, 0xf5, 0xd9, 0xd9, 0xd8, 0x9f, 0x59, 0x9e, 0xed, 0x56, 0xb9, 0x51, 0x86, 0x65, 0xd9, 0x2e, 0xd5, 0x70, 0x4a, 0xb0, 0x1c, 0xdb, 0x9d, 0x1a, 0x4e, 0x9, 0x96, 0x63, 0xbb, 0x53, 0xc7, 0x8a, 0x63, 0x19, 0xb6,
-0x2b, 0xb5, 0xbc, 0x28, 0x96, 0x5f, 0xbb, 0x51, 0xcb, 0x8b, 0x62, 0xf9, 0xb5, 0x1b, 0xf5, 0xcc, 0x18, 0x96, 0x5d, 0xbb, 0x50, 0xd3, 0xcd, 0xba, 0xc7, 0xda, 0xb1, 0xdc, 0xac, 0x7b, 0xac, 0x1d, 0x46, 0xbb, 0x50, 0xef, 0x1e, 0x6b, 0x87, 0xd2, 0x6e, 0xd4, 0xba, 0xc7, 0xda, 0xa1, 0xb4, 0x1b, 0xb5, 0xee, 0xb1, 0x76, 0x38, 0xed, 0x4a, 0x9d, 0x7b, 0xac, 0x9d, 0x84, 0x76, 0xa7, 0xc6, 0x3d, 0xd6, 0x4e, 0x42, 0xbb, 0x53, 0xe3, 0x1e, 0x6b, 0x27, 0xa3, 0x5d, 0x2a, 0xbf, 0xc7, 0xda, 0x49,
-0x69, 0xb7, 0x4a, 0xef, 0xb1, 0x76, 0x72, 0xda, 0xb5, 0xb2, 0x7b, 0xac, 0x9d, 0x9c, 0x76, 0xad, 0xec, 0x1e, 0x6b, 0xa7, 0x41, 0xbb, 0x57, 0x72, 0x8f, 0xb5, 0xd3, 0xa1, 0x5d, 0x8c, 0xdf, 0x63, 0xed, 0x74, 0x68, 0x17, 0xe3, 0xf7, 0x58, 0x3b, 0x2d, 0xda, 0xcd, 0xe8, 0x3d, 0xd6, 0x4e, 0x8f, 0x76, 0xb5, 0xd5, 0x8d, 0x62, 0x59, 0xb4, 0x4b, 0x18, 0xce, 0xab, 0x1b, 0xb5, 0x5e, 0xdd, 0x56, 0x37, 0xce, 0xea, 0x46, 0x59, 0xdd, 0x18, 0xab, 0x1b, 0x65, 0x75, 0xa3, 0xac, 0x6e, 0x8c, 0xd5,
-0x8d, 0xb2, 0xba, 0x51, 0x56, 0x37, 0xc6, 0xea, 0x46, 0x59, 0xdd, 0x28, 0xab, 0x1b, 0x63, 0x75, 0xa3, 0xac, 0x6e, 0x94, 0xd5, 0x8d, 0xb1, 0xba, 0x51, 0x56, 0x37, 0x4a, 0xbb, 0xdb, 0x2f, 0x1d, 0x7d, 0x65, 0x6e
-};
-
-const uint8_t center_1[383] PROGMEM = {
-0x78, 0x9c, 0xed, 0xd2, 0x81, 0xd, 0xc2, 0x30, 0x10, 0x4, 0xc1, 0x74, 0xee, 0xd2, 0x83, 0x44, 0x5, 0x64, 0xe5, 0x60, 0x88, 0x67, 0xaf, 0x81, 0xd7, 0xe8, 0x8f, 0xe3, 0xb7, 0x3a, 0xc7, 0x7b, 0x57, 0x1a, 0xab, 0x6f, 0x5e, 0xd9, 0x65, 0xad, 0xad, 0xdd, 0xc2, 0x77, 0x6d, 0xed, 0x36, 0xc5, 0x6b, 0x23, 0xb7, 0x89, 0x5a, 0x1b, 0xb8, 0x4d, 0xfd, 0xae, 0xd, 0xdc, 0x6e, 0xd5, 0x7a, 0x9c, 0xdb, 0xcd, 0xdf, 0xf5, 0x38, 0xb7, 0xaf, 0x7b, 0x3d, 0xc5, 0x6d, 0x55, 0xdc, 0xb8, 0x71, 0xfb,
-0x38, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0x2d, 0x6e, 0xab, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2,
-0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2,
-0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2, 0xd6, 0xe2,
-0xd6, 0xda, 0xd1, 0xed, 0x1c, 0x13, 0x16, 0xdd, 0x66, 0x6c, 0x99, 0xdb, 0x7f, 0xc7, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0xc5, 0xad, 0x75, 0xd9, 0xed, 0x5, 0xc9, 0xe3, 0x64, 0xec
-};
-
-
-const uint8_t center_2[673] PROGMEM = {
-0x78, 0x9c, 0xed, 0xdc, 0x51, 0x42, 0xeb, 0x30, 0xc, 0x5, 0x51, 0x76, 0xde, 0xa5, 0x17, 0x42, 0xe1, 0xb, 0xa, 0xcd, 0x58, 0xb6, 0xa4, 0x78, 0x86, 0x5, 0x70, 0x73, 0x12, 0xe8, 0x7b, 0x34, 0xe9, 0xdb, 0x5b, 0xad, 0xee, 0xb7, 0xdf, 0xbf, 0xb2, 0x77, 0xd5, 0xeb, 0x4b, 0xe6, 0xd5, 0xb6, 0x57, 0x3c, 0xa5, 0xf5, 0x44, 0x30, 0xfb, 0x18, 0xd6, 0x36, 0xe8, 0xb5, 0xa1, 0xdf, 0xf0, 0x35, 0xf6, 0x87, 0x5e, 0xf6, 0xb1, 0xcd, 0x6a, 0x9a, 0xd8, 0x85, 0xed, 0x16, 0x98, 0x5d, 0xce, 0x6e,
-0xa1, 0xd9, 0x85, 0xec, 0x12, 0xd4, 0xda, 0xdb, 0xa5, 0x99, 0xb5, 0x96, 0x4b, 0x56, 0x6b, 0x2a, 0x57, 0x40, 0xed, 0x51, 0x2b, 0xbb, 0x32, 0x6a, 0x47, 0x4d, 0xe4, 0x4a, 0x99, 0x3d, 0x6a, 0x20, 0x57, 0x50, 0xed, 0xa8, 0xb8, 0x5c, 0x51, 0xb5, 0xa3, 0xd2, 0x72, 0xd9, 0x38, 0x7f, 0x56, 0x56, 0xae, 0xf0, 0xd5, 0xf6, 0xa8, 0xa4, 0x5c, 0x79, 0xb5, 0x8f, 0xb2, 0x8d, 0x7e, 0xd6, 0x41, 0xed, 0x5e, 0xee, 0x8a, 0x6b, 0xa2, 0x76, 0x54, 0x4a, 0x2e, 0x1b, 0xe3, 0x54, 0x65, 0xe4, 0x1a, 0x5d,
-0x6d, 0x9f, 0x65, 0x7b, 0x7d, 0x97, 0xed, 0x70, 0xba, 0x12, 0x57, 0x5c, 0xb7, 0xab, 0xed, 0x28, 0xdb, 0x2c, 0x58, 0xed, 0xe9, 0x7b, 0xa3, 0xe1, 0xef, 0xe7, 0xa4, 0x5f, 0x71, 0x41, 0xc7, 0xf3, 0xd2, 0x7b, 0x79, 0x91, 0x7a, 0xf3, 0x65, 0xfe, 0x39, 0x96, 0xf1, 0x4e, 0x9e, 0xfb, 0xa8, 0x33, 0x35, 0xc7, 0x63, 0xd1, 0x31, 0xc0, 0xf5, 0x11, 0x76, 0xb1, 0x12, 0x27, 0xf7, 0xa7, 0xa8, 0x7d, 0x7e, 0xe7, 0x71, 0xb9, 0xb4, 0x2b, 0x6e, 0xf4, 0xfe, 0x8e, 0xee, 0xdf, 0x3f, 0x65, 0x77, 0xc8,
-0xea, 0x31, 0xb9, 0x88, 0x5, 0x68, 0x75, 0x81, 0xc5, 0xf9, 0xe7, 0x6e, 0xe1, 0xe2, 0xd0, 0xbd, 0xdb, 0xb8, 0x5, 0xaf, 0xad, 0x72, 0xfe, 0x5e, 0xde, 0xb, 0x9b, 0xb0, 0x4, 0xcb, 0xc5, 0x6f, 0x79, 0x61, 0x2d, 0x6b, 0xc2, 0x39, 0xee, 0x74, 0xc5, 0xd1, 0xad, 0xb5, 0xd6, 0xb4, 0x71, 0x9b, 0xb6, 0x53, 0x37, 0xb8, 0xa7, 0x8b, 0x1b, 0x6a, 0xe2, 0x9e, 0x5b, 0xb5, 0x45, 0x4f, 0x76, 0x92, 0x26, 0x9e, 0x5d, 0xdd, 0x96, 0x2e, 0x5a, 0xec, 0xc6, 0xce, 0xee, 0x7e, 0x9b, 0x3a, 0x6c, 0xac,
-0xb8, 0x29, 0x64, 0xe3, 0xe4, 0xd7, 0x2e, 0xdd, 0x16, 0x6e, 0x5a, 0xfc, 0x2f, 0x11, 0xdd, 0x16, 0x6e, 0xd4, 0xed, 0x42, 0x6e, 0x73, 0x37, 0x85, 0x6c, 0xd4, 0x4d, 0xb7, 0x95, 0x1b, 0x75, 0xab, 0xe9, 0x86, 0x9a, 0xbb, 0xe9, 0xc7, 0x46, 0xdd, 0xd8, 0x46, 0xe2, 0x36, 0xfb, 0xff, 0x59, 0xba, 0x2d, 0x5b, 0x94, 0xf2, 0xf7, 0xde, 0xd3, 0x5f, 0x93, 0xf7, 0xb4, 0x70, 0xab, 0x16, 0x52, 0xdb, 0xde, 0xad, 0xcb, 0x5f, 0x7b, 0xab, 0xa5, 0x1b, 0xa9, 0xd7, 0xfb, 0xf5, 0x75, 0xea, 0xf3, 0xae,
-0x73, 0xa5, 0x3a, 0xdd, 0xe5, 0x50, 0x29, 0x7f, 0x4a, 0x49, 0x5e, 0x6d, 0x2c, 0xac, 0xb6, 0xb5, 0xdb, 0xc8, 0x5d, 0xaa, 0xd9, 0xdb, 0xf3, 0xea, 0x77, 0x6f, 0x6f, 0x8d, 0x6, 0xd4, 0x36, 0x76, 0x53, 0x8d, 0xd4, 0xf5, 0x99, 0x8f, 0xdc, 0x46, 0x9f, 0x32, 0xca, 0xde, 0x9f, 0x53, 0xd6, 0xd3, 0x74, 0xbd, 0x53, 0x8d, 0xd4, 0xf9, 0x39, 0xc0, 0xbc, 0x54, 0x23, 0x75, 0x7f, 0x5a, 0x37, 0xa7, 0x90, 0xe7, 0xc3, 0xb7, 0xbb, 0xda, 0x54, 0x23, 0xc5, 0x7c, 0x16, 0x41, 0xf6, 0x51, 0xac, 0xee,
-0xa, 0x9f, 0xe0, 0xb0, 0x3e, 0xd5, 0x48, 0xaa, 0x91, 0x42, 0xcc, 0xee, 0x7b, 0xfd, 0x66, 0xb, 0xfc, 0x44, 0xa4, 0x8d, 0xae, 0x36, 0xd5, 0x48, 0xaa, 0x91, 0x54, 0x23, 0xa9, 0x46, 0x52, 0x8d, 0xa4, 0x1a, 0x49, 0x35, 0x92, 0x6a, 0x24, 0xd5, 0x48, 0xaa, 0x91, 0x54, 0x23, 0xa9, 0x46, 0x52, 0x8d, 0x14, 0xa6, 0xb6, 0x91, 0x99, 0x6a, 0x2c, 0xd5, 0x48, 0xaa, 0x91, 0x54, 0x23, 0x5, 0x99, 0xa9, 0xa6, 0xda, 0x3f, 0xf9, 0xf3, 0x49, 0x52, 0x8d, 0xa4, 0x1a, 0x49, 0x35, 0x92, 0x6a, 0x24,
-0xd5, 0x48, 0xaa, 0x91, 0x54, 0x23, 0x79, 0x17, 0x16, 0x49, 0x35, 0x92, 0x6a, 0x24, 0xd5, 0x48, 0xaa, 0x91, 0x54, 0x23, 0xa9, 0x46, 0x52, 0x8d, 0xe4, 0xd3, 0x40, 0x24, 0xd5, 0x48, 0xaa, 0x91, 0x2, 0xcc, 0xb6, 0x53, 0xf3, 0xb5, 0x80, 0xa4, 0x1a, 0x49, 0x35, 0x92, 0xaf, 0x5, 0x24, 0xd5, 0x48, 0x81, 0x77, 0xfc, 0xa5, 0xa7, 0x1a, 0x6b, 0xa1, 0xdb, 0xa5, 0xd2, 0x8d, 0xa5, 0x1b, 0x4b, 0x37, 0x96, 0x6e, 0x2c, 0xdd, 0x58, 0xba, 0xb1, 0x74, 0x63, 0xe9, 0xc6, 0xd2, 0x8d, 0xa5, 0x1b,
-0x4b, 0x37, 0x96, 0x6e, 0x2c, 0xdd, 0x58, 0xba, 0xb1, 0x74, 0x63, 0xe9, 0xc6, 0xd2, 0x8d, 0xa5, 0x1b, 0x4b, 0x37, 0x96, 0x6e, 0x2c, 0xdd, 0x58, 0xeb, 0xdc, 0x76, 0xef, 0x1d, 0x2c, 0xb6, 0x57, 0xab
-};
-
-const uint8_t center_3[893]PROGMEM = {
-0x78, 0x9c, 0xed, 0x9d, 0x51, 0x6e, 0xe3, 0x30, 0xc, 0x5, 0x7b, 0xf3, 0x1e, 0x3d, 0xbb, 0x82, 0x9b, 0x45, 0xbb, 0x80, 0x53, 0x9b, 0x22, 0xf9, 0x86, 0xc, 0xa7, 0xbf, 0x5, 0x32, 0x1a, 0xcb, 0x6e, 0x1a, 0xc9, 0xce, 0xc7, 0xc7, 0x50, 0x9f, 0xc7, 0xe7, 0xab, 0x1f, 0xb5, 0x1d, 0x8b, 0xaf, 0x2a, 0xd7, 0x79, 0xf3, 0x86, 0x37, 0x6b, 0x9d, 0x14, 0x54, 0x8f, 0x22, 0x13, 0x87, 0x62, 0xff, 0xf5, 0x53, 0x8f, 0x28, 0x1e, 0xf7, 0x66, 0xff, 0xda, 0xf5, 0xad, 0x17, 0xd6, 0xac, 0x71, 0xbb,
-0xf0, 0x66, 0xdf, 0xea, 0xa9, 0xc7, 0xea, 0x45, 0x62, 0xb3, 0x46, 0xed, 0xd2, 0x9b, 0xb5, 0x28, 0x27, 0xaa, 0x56, 0xba, 0x9d, 0xb4, 0x59, 0xd9, 0x72, 0x80, 0x6a, 0x8b, 0x62, 0xe5, 0x20, 0xd5, 0xe, 0xca, 0xb4, 0x43, 0x55, 0xfb, 0x8b, 0xba, 0xc7, 0x35, 0x68, 0xd5, 0x1e, 0x25, 0x66, 0x1c, 0xb0, 0xda, 0x2, 0x5e, 0xe, 0x5a, 0x6d, 0x1, 0x2e, 0x7, 0xae, 0xb6, 0x80, 0x96, 0x83, 0x57, 0x5b, 0x0, 0xcb, 0x15, 0xa8, 0xf6, 0x20, 0xfe, 0x65, 0x55, 0x17, 0xb9, 0x8, 0x6c, 0xc6, 0xd5,
-0x98, 0x6d, 0xb, 0x75, 0xa9, 0x9f, 0xa8, 0x6b, 0xdc, 0x0, 0x34, 0xe3, 0x42, 0x66, 0x5b, 0xdc, 0xa7, 0x76, 0x98, 0x72, 0x5e, 0xe3, 0x39, 0x1f, 0x91, 0x6b, 0x45, 0x48, 0x37, 0x87, 0x11, 0x5d, 0x5e, 0x19, 0x70, 0xea, 0x7, 0x28, 0x97, 0xbf, 0xa, 0xea, 0xb1, 0xf2, 0x1a, 0xd3, 0xe2, 0xe6, 0x28, 0x4, 0x23, 0xd8, 0x2d, 0xe7, 0xdb, 0x20, 0xd9, 0x7f, 0xf3, 0xa8, 0x2b, 0x5f, 0x7b, 0x97, 0xd, 0x77, 0x7, 0x73, 0xed, 0xab, 0x6f, 0x99, 0x8b, 0xbd, 0xed, 0xe5, 0x7c, 0x5e, 0x3f, 0xd9,
-0xda, 0xf1, 0x68, 0x13, 0x1c, 0xb2, 0x9c, 0x11, 0x16, 0xca, 0x6e, 0x36, 0x9c, 0x8d, 0x19, 0x47, 0x2f, 0xdc, 0x37, 0xe0, 0x38, 0xdb, 0x4c, 0xfc, 0x3d, 0xaa, 0xd9, 0x72, 0x4c, 0x2e, 0xd9, 0x5a, 0x8, 0xba, 0xaa, 0x98, 0xca, 0x89, 0xae, 0x70, 0x96, 0x6c, 0x61, 0x2e, 0xbd, 0xbb, 0x5, 0x9a, 0xb2, 0x6c, 0x5e, 0x78, 0x7e, 0xb2, 0x4c, 0x59, 0x36, 0x2f, 0x3c, 0xd, 0xdd, 0x60, 0x3e, 0xd3, 0xcd, 0xe6, 0x53, 0xa5, 0x5b, 0xa8, 0x27, 0xed, 0x38, 0x9e, 0x7a, 0xde, 0x67, 0xba, 0x7d, 0xf0,
-0xba, 0x19, 0x8c, 0x34, 0xe7, 0x29, 0xac, 0x1b, 0xed, 0xba, 0xf1, 0xc2, 0xf3, 0xe6, 0x4f, 0xb8, 0x4f, 0x89, 0x6e, 0x34, 0xa6, 0x9b, 0x8d, 0xe9, 0x66, 0x63, 0xba, 0xd9, 0xb8, 0x5d, 0x4d, 0xbe, 0x16, 0x48, 0xa0, 0xca, 0xbb, 0x37, 0x1a, 0x73, 0x96, 0x5a, 0xa0, 0x7d, 0x36, 0x53, 0x85, 0x39, 0x4b, 0x2d, 0xcc, 0x6c, 0xb3, 0xc0, 0x59, 0x55, 0xab, 0x44, 0xb5, 0xb5, 0x53, 0xa, 0xb6, 0x6a, 0xef, 0x3d, 0xdb, 0x6a, 0xee, 0xa9, 0x51, 0x53, 0x77, 0xf, 0x97, 0x8e, 0xbd, 0xdd, 0x96, 0x6a,
-0x7b, 0x5, 0x3d, 0xf6, 0xf6, 0x66, 0xd2, 0x67, 0x27, 0xf9, 0x3e, 0x17, 0x3f, 0x29, 0x76, 0x44, 0x3d, 0x62, 0x1f, 0x3c, 0x8b, 0x5c, 0xa2, 0xc5, 0x6c, 0x4b, 0xef, 0xd6, 0xa4, 0x5a, 0x72, 0xb7, 0x36, 0xd5, 0x52, 0xbb, 0x35, 0xaa, 0x96, 0xd8, 0xad, 0x55, 0xb5, 0xa4, 0x6e, 0xcd, 0x9a, 0x2d, 0xa6, 0x9a, 0x8d, 0x69, 0x66, 0x63, 0x9a, 0xd9, 0x98, 0x66, 0x36, 0x82, 0xba, 0xb5, 0x2f, 0x17, 0xd6, 0xad, 0x79, 0xbb, 0xd0, 0x6e, 0x8d, 0xdb, 0x85, 0x77, 0x6b, 0xda, 0x2e, 0xa5, 0x5b, 0xc3,
-0x72, 0x49, 0xdd, 0xda, 0xb5, 0x4b, 0xec, 0xd6, 0xaa, 0x5c, 0x6a, 0xb7, 0x46, 0xe5, 0x92, 0xbb, 0xb5, 0x29, 0x97, 0xde, 0xad, 0x49, 0x39, 0x41, 0xb7, 0x16, 0xe5, 0x72, 0xd7, 0xb2, 0x3a, 0x95, 0xbb, 0x8a, 0x6f, 0x45, 0xf5, 0x68, 0xf2, 0x99, 0xd5, 0xe7, 0x1d, 0xf6, 0xdb, 0xa9, 0x47, 0xa0, 0x43, 0xf3, 0xfc, 0xb4, 0xe, 0x6c, 0x9e, 0xb3, 0x6f, 0x5c, 0x6e, 0x76, 0xc1, 0xd9, 0x99, 0x5d, 0x97, 0x56, 0x2a, 0x3e, 0xc9, 0x8c, 0xc1, 0xec, 0xc5, 0xb7, 0x31, 0xbb, 0xf1, 0xad, 0xcc, 0x8c,
-0xb3, 0x32, 0xdd, 0x6c, 0xcc, 0x5d, 0x5a, 0x36, 0x6a, 0x3d, 0x3b, 0x8f, 0xc4, 0x74, 0xb3, 0x31, 0x67, 0xaa, 0x8d, 0xe9, 0x66, 0x63, 0xba, 0x59, 0x99, 0x6e, 0x36, 0xd8, 0xdd, 0xb8, 0x76, 0x5c, 0x33, 0xb6, 0x1d, 0xd7, 0x8c, 0x6d, 0xc7, 0x35, 0x33, 0xda, 0x25, 0xbd, 0xbf, 0x84, 0x77, 0xfb, 0x64, 0x76, 0xa3, 0xbf, 0xf, 0x99, 0x6e, 0xbd, 0xfc, 0xa8, 0x5e, 0x74, 0x3f, 0x83, 0x55, 0xf2, 0xff, 0xf5, 0x4c, 0x43, 0xa6, 0xd5, 0xae, 0x21, 0xf0, 0x79, 0xaa, 0xf9, 0xdd, 0x80, 0x8e, 0xd4,
-0xab, 0xc7, 0xb6, 0x63, 0xa8, 0x25, 0xcf, 0xe8, 0xc4, 0xd3, 0x44, 0xa0, 0x4f, 0xeb, 0x6e, 0xa8, 0x6f, 0x97, 0x79, 0x28, 0x3e, 0x25, 0x67, 0xed, 0x2d, 0x20, 0xb9, 0xfc, 0x62, 0x6a, 0xed, 0x16, 0x70, 0x8c, 0x59, 0xc7, 0xf0, 0x57, 0x5b, 0x2b, 0x90, 0x6f, 0x6b, 0x53, 0xad, 0x65, 0x31, 0x76, 0x9e, 0x31, 0x2c, 0x6e, 0x3a, 0xef, 0xe0, 0x62, 0x5d, 0x73, 0xbf, 0xa5, 0x76, 0x87, 0xed, 0xee, 0xde, 0x68, 0x9f, 0x6, 0xa, 0xf3, 0x9d, 0x72, 0xb5, 0x9f, 0x9e, 0xe7, 0x70, 0x2f, 0x81, 0xc1,
-0xdf, 0xe3, 0xe, 0x6, 0xff, 0x16, 0x37, 0xc7, 0xe0, 0xc1, 0xe5, 0xef, 0x50, 0x71, 0xbb, 0x67, 0x46, 0xbe, 0x6a, 0xea, 0x7a, 0x7, 0xd5, 0x69, 0xbf, 0x8e, 0x4f, 0x1d, 0xf4, 0x1c, 0x4f, 0x12, 0xf2, 0xd9, 0xb6, 0x8, 0xba, 0x67, 0x2f, 0xe, 0x44, 0xb5, 0x45, 0xb1, 0x72, 0xd3, 0xcd, 0x2, 0xa6, 0xda, 0xa2, 0x4c, 0x39, 0x54, 0xb5, 0x45, 0x89, 0x72, 0xb8, 0x6a, 0xb, 0x7c, 0x39, 0x64, 0xb5, 0x5, 0xbb, 0x9c, 0xba, 0xce, 0x2b, 0xd4, 0x6d, 0x5e, 0x80, 0x9d, 0x6d, 0x7, 0xea, 0x3c,
-0x27, 0xc0, 0xab, 0x41, 0xcf, 0x55, 0x7c, 0xb5, 0x3, 0x54, 0xbb, 0x22, 0xcd, 0xe, 0x30, 0xe5, 0x4a, 0x55, 0x5b, 0x20, 0xca, 0x95, 0xab, 0x76, 0x20, 0x6d, 0x57, 0xb4, 0xd9, 0x81, 0xac, 0x5c, 0xe9, 0x6a, 0x7, 0xe9, 0xed, 0x1a, 0x34, 0x7b, 0x92, 0xd6, 0xae, 0x51, 0xb3, 0x27, 0xe1, 0xed, 0x1a, 0x36, 0x7b, 0x12, 0xd6, 0xae, 0x71, 0xb3, 0x27, 0xae, 0x6b, 0x2c, 0x97, 0x57, 0xc2, 0xfa, 0xb0, 0xfb, 0x54, 0xa8, 0xf7, 0x2b, 0xf6, 0x93, 0x1b, 0x4f, 0x64, 0xfc, 0xfa, 0x4d, 0xb5, 0x31,
-0x95, 0xef, 0xcf, 0xb7, 0x54, 0xbb, 0x9c, 0xf1, 0x7, 0x1f, 0x70, 0xc, 0xd
-};
