--- steering/Reference/steering-display/src/tft_data.h
@@ -1,10 +1,0 @@
-
-#ifndef TFT_DATA_H
-#define TFT_DATA_H
-
-#if !defined(PROGMEM)
-	#define PROGMEM
-#endif
-
-extern const uint8_t logo[5298] PROGMEM, center_n[1075] PROGMEM, center_1[383] PROGMEM, center_2[673] PROGMEM, center_3[893] PROGMEM;
-#endif /* TFT_DATA_H */
