--- steering/display/EVE_Test_Arduino_PlatformIO/.gitignore
@@ -1,2 +1,0 @@
-.pio
-.vscode
