--- steering/display/EVE_Test_Arduino_PlatformIO/readme.md
@@ -1,6 +1,0 @@
-# EVE_Test_Arduino_PlatformIO Example
-
-This example project uses the [FT800-FT813 library by Rudolph Riedel.](https://github.com/RudolphRiedel/FT800-FT813/) It is a modified version of the example in that repository, configured for the Q23 steering display. 
-
-For a template with empty splash screen
-
