+++ steering/src/display/displaySplash.cpp
